% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cala.R
\name{cala}
\alias{cala}
\title{Calibration of Independently Estimated Sets of Item Parameters
by Minimizing the LS Criterion Defined in terms of Parameter Values}
\usage{
cala(
  indata,
  baseform = 0,
  nsubj = 1,
  maxiter = 100,
  eps = 0.00001,
  init = 1,
  print = 1,
  debug = 0
)
}
\arguments{
\item{indata}{data frame containing at least following: \cr
           giid  gfid  type ncat  a  p1 p2 .... \cr
indata can be created from a parameter data frame
by renaming name as giid and adding gfid.
It is assumed that indata is sorted by gfid.}

\item{baseform}{form number or 0}

\item{nsubj}{a vector consisting of the # of subjects in each form or 1.
\cr To be used when calculating the mean of theta of the combined group.}

\item{maxiter}{max of iteration for PCA}

\item{eps}{eos for PCA}

\item{init}{not used}

\item{print}{> 0 to print results}

\item{debug}{= 1 to print intermediate result}
}
\value{
A list of: \cr
 param Data frame containing the equated item parameters \cr
 uv  Transformations from common scale to each scale\cr
 qr  Transformations from each scale to common scale\cr
 nfini Vector of # of forms including each item \cr
 niinf Vector of # of items included in each form \cr
 fini  List of the Forms including each item. \cr
 iinf  List of the Items included in each form.\cr
 iftable  Matrix of item x form indicating the inclusion pattern.\cr
 toP  Vector of stating locations in ParamHat \cr
 fromP  Vector of ending locations in ParamHat  \cr
 iterPCA  # of iterations needed for convergence. \cr
 nbPCA  3 of b-parameters processed by PCA.\cr
 nisolo # of items that are included in only one form. \cr
 ParamHat Item Paramter matrix analyzed by PCA \cr
 rmse LS criterion minimized.
}
\description{
Calibration of Independently Estimated Sets of Item Parameters
by Minimizing the LS Criterion Defined in terms of Parameter Values
}
\details{
This program finds u,v and b such taht \cr
     bhata[j,g] = u[g] + b[j] v[g] + error \cr
     j=1,2,..., # of items and  g=1,2,..., # of forms, \cr
  using PCA with missing data, where bhat is the collection of estimated
  b-parameters. \cr
 \cr
   a[g] will be estimated as the average of ahat[j,g]/r[g] , \cr
   c[g] will be estimated as the average of chat[j,g] , \cr
  where r[g]=1/v[g]
}
\examples{
# Binary Items
res=cala( indata=paramCal1, baseform=1, print=2  )

# Mixture of Item Types
res=cala( indata=paramCal2, baseform=1, print=2  )

}
\references{
Sayaka Arai & Shin-ichi Mayekawa (2011)
A comparison of equating methods and linking designs for developing
an item pool under item response theory.
Behaviormetrika, Vol. 38 No. 1, pp. 1-16 \cr

Shin-ich Mayekawa (1991) Chapter 4. Parameter Estimation.
In Shiba Ed. Item Response Theory. pp87-129. (in Japanese)
}
