% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calr.R
\name{calr}
\alias{calr}
\title{Calibration of Independently Estimated Sets of Item Parameters}
\usage{
calr(
  indata,
  baseform = 0,
  nsubj = 0,
  npoints = 21,
  thmin = -3,
  thmax = 3,
  maxiter = 100,
  eps = 0.00001,
  print = 1,
  debug = 0
)
}
\arguments{
\item{indata}{data frame containing at least following: \cr
           giid  gfid  type ncat  a  p1 p2 .... \cr
indata can be created from a parameter data frame
by renaming name as giid and adding gfid.
It is assumed that indata is sorted by gfid.}

\item{baseform}{form number or 0}

\item{nsubj}{a vector consisting of the # of subjects in each form or 1.
\cr To be used when calculating the mean of theta of the combined group.}

\item{npoints}{# of discrete theta points to be used.}

\item{thmin}{Minimum value of theta points.}

\item{thmax}{Maximum value of theta points.}

\item{maxiter}{max of iterations}

\item{eps}{Criterion of convergence for the relative improvement of
rss value.}

\item{print}{> 0 to print results}

\item{debug}{= 1 to print intermediate result}
}
\value{
A list of: \cr
 param Data frame containing the equated item parameters \cr
 uv  Transformations from common scale to each scale\cr
 qr  Transformations from each scale to common scale\cr
 nfini Vector of # of forms including each item \cr
 niinf Vector of # of items included in each form \cr
 fini  List of the Forms including each item. \cr
 iinf  List of the Items included in each form.\cr
 iftable  Matrix of item x form indicating the inclusion pattern.\cr
 iter # of iteration required for convergence \cr
 rss # LS criterion minimized \cr
 rssimpr Final relative improvement of rss. \cr
 eps Convergence criterion for rssimpr \cr
 rmse rmse calculated from rss \cr
 nisolo # of items that are included in only one form. \cr
 gradI Final gradients of common item parameters. \cr
 gradF Final gredients or transformation constants. \cr
}
\description{
This function calibrates (equates) independently estimated sets of item
parameters (with common linking items) by minimizing the LS criterion 
defined in terms of item response functions (irf) \cr
Japanese help file: (\link[lazy.irt]{calr_JPH})
}
\details{
This program finds q, r and item param such taht \cr
    P(theta_g | est_param_g) = P( q[g] + r[g] theta | param )  + error \cr
     j=1,2,..., # of items and  g=1,2,..., # of forms, \cr
 where param is the item parameters on the common scale. \cr
\cr
 First, those items which are included in at least two forms will be
 picked up and analyzed. \cr
 Then, the item parameters for the solo items will be converted
 using estimated q and r. \cr
\cr
 Subscript j is for item, g for form, q is for theta point.  \cr
}
\examples{
# Binary Items
res=calr( indata=paramCal1, baseform=1, print=2  )

# Mixture of Item Types
res=calr( indata=paramCal2, baseform=1, print=2  )

}
\references{
Sayaka Arai, Tomohiro Ohtani, & Shin-ichi Mayekawa (2013)
Simultaneous Equating of Separately Calibrated Polytomous IRT Test Items.
IMPS 2013, CITO, Arnhem, The Netherland. 23-26 July 2013 \cr

Shin-ichi Mayekawa (2012)
Simultaneous equating of separately calibrated item parameters
under the common item design.
COMPSTAT 2012, Amathus Beach Hotel, Limassol, Cyprus. 27-31 August 2012\cr

Sayaka Arai & Shin-ichi Mayekawa (2011)
A comparison of equating methods and linking designs for developing
an item pool under item response theory.
Behaviormetrika, Vol. 38 No. 1, pp. 1-16 \cr

Shin-ich Mayekawa (1991) Chapter 4. Parameter Estimation.
In Shiba Ed. Item Response Theory. pp87-129. (in Japanese)
}
