% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calr_JPH.R
\docType{package}
\name{calr_JPH}
\alias{calr_JPH}
\title{独立に推定された項目パラメタの共通項目法による等化}
\usage{
calr(indata, baseform = 0, nsubj = 0
, npoints = 21, thmin = -3, thmax = 3
, maxiter = 100, eps = 0.00001, print = 1, debug = 0)
}
\arguments{
\item{indata}{以下の変数を持つ項目パラメタのデータフレーム: \cr
           \code{giid  gfid  type ncat  a  p1 p2 ....} \cr
ただし、 \code{giid} は項目名、\code{gfid} は項目パラメタが推定された
集団名で、\code{type, ncat, p1, p2, ...} 等は項目パラメタデータフレーム
の説明 \link[lazy.irt]{lazy.irt_JPH} やテストデータを参照のこと。}

\item{baseform}{基準となる集団名または 0}

\item{nsubj}{各集団の受験者数または 1（重みとして使われる）}

\item{npoints}{特性値 theta の離散点の数}

\item{thmin}{theta の最小値}

\item{thmax}{theta の最大値}

\item{maxiter}{最大繰り返し数}

\item{eps}{収束判定に用いられる rss の最大変化率}

\item{print}{> 0 結果を出力}

\item{debug}{= 1 途中結果を出力}
}
\value{
以下の要素からなるリスト
\preformatted{
 param 推定された共通尺度上の項目パラメタデータフレーム
 uv  共通尺度から個別尺度への変換係数
 qr  共通尺度から個別尺度への変換係数
 nfini 各項目を受験している集団の数からなるベクトル
 niinf 各集団が受験している項目の数からなるベクトル
 fini  各項目を受験している集団のリスト
 iinf  各集団が受験している項目のリスト
 iftable  項目 x 集団の項目と集団の包含関係を示す行列で
 iter 収束までのゝ計算数
 rss 最小化された RSS の値
 rssimpr 最後の RSS の減少率
 eps 収束判定基準
 rmse rss から計算された rmse
 nisolo 一つの集団のみで受験された項目の数
 gradI RSS の項目パラメタに関する１次微分
 gradF RSS の変換係数に関する１次微分
 }
}
\description{
この関数は、共通項目を含む複数のテスト冊子が複数の能力が異なる集団で
実施された場合、集団ごとに推定された項目パラメタの値を元に、これらの
項目パラメタの値を共通尺度上で等化することを目的としている。 \cr
English help file: (\link[lazy.irt]{calr})
}
\details{
この関数は以下の最小２乗基準を最小とする変換係数ベクトル \code{q} と
 \code{r} ならびに項目パラメタのセット \code{param} を求める。\cr
 \eqn{ RSS= \sum_g\sum_j \delta_{gj} \sum_k \sum_a
 ( P_{kj}(\theta_{a}|param_{jg}) - P_{kj}(q_g + r_g \theta_{a}|param_j) )^2
 \; h(\theta_a) }
 \cr
 ただし、
 \eqn{\theta_a} は区間 \code{[thmin,thmax]} を \code{npoints} の
 離散点に分割したもの、\eqn{h(\theta_a)} は \eqn{\theta_a} における
 標準正規分布の密度関数に比例する重み、
 \eqn{param_{jg}} は第 \code{g} 群で推定された項目 \code{j}
  の項目パラメタ、\eqn{param_{j}} は共通尺度上の項目 \code{j} の
  項目パラメタ、\code{j=1,2,..., 総項目数} \code{g=1,2,..., 総集団数} であり
 \eqn{\delta_{gj}} は集団 \code{g} が項目 \code{j} を受験しているときに
 1 となる指標関数である。#'
 したがって、\code{RSS} の中に現れる \eqn{P_{kj}(\theta_{a}|param_{jg})}
  は定数であり未知パラメタは含まれないことに注意。

 なお、各群ごとの項目パラメタは、各群の特性値の分布の平均が 0 で標準偏差が
 1 であることを仮定して推定されたものとする。

 上記のパラメタが求まった後に、\code{baseform} の指定に従い、
 共通尺度をどの集団に設定するか、もしくは全集団における特性値の分布の
 平均と分散を 0 と 1 に合わせるか（\code{baseform=0}）の処理を行う。
}
\examples{
# Binary Items
res=calr( indata=paramCal1, baseform=1, print=2  )

# Mixture of Item Types
res=calr( indata=paramCal2, baseform=1, print=2  )



# @keywords internal
}
\references{
Sayaka Arai, Tomohiro Ohtani, & Shin-ichi Mayekawa (2013)
Simultaneous Equating of Separately Calibrated Polytomous IRT Test Items.
IMPS 2013, CITO, Arnhem, The Netherland. 23-26 July 2013 \cr

Shin-ichi Mayekawa (2012)
Simultaneous equating of separately calibrated item parameters
under the common item design.
COMPSTAT 2012, Amathus Beach Hotel, Limassol, Cyprus. 27-31 August 2012\cr

Sayaka Arai & Shin-ichi Mayekawa (2011)
A comparison of equating methods and linking designs for developing
an item pool under item response theory.
Behaviormetrika, Vol. 38 No. 1, pp. 1-16 \cr

前川眞一 (1991) 第４章　パラメタの推定.
芝（編）項目反応理論. pp87-129.
}
