% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkparam.R
\name{checkparam}
\alias{checkparam}
\title{Checking and Subsetting the Parameter Data Frame}
\usage{
checkparam(
  param,
  type = NULL,
  modulename = "checkparam",
  tomat = 1,
  printerror = 1
)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{type}{Item type}

\item{modulename}{name of the module which calls this function}

\item{tomat}{= 0 not to convert the result to a matrix.}

\item{printerror}{= 0 to suppress error messages}
}
\value{
A parameter matrix or data frame.
}
\description{
Checking and Subsetting the Parameter Data Frame
}
\details{
This function checks if the input \code{param} is a valid parameter
data frame with \code{name, type, ncat, p1, p2, p3, ...,}  variables.\cr
When \code{type = (B | B3 | Bn | Bn3 | G | Gn | P |  PN | N)} is given,
the items with the given type will be selected
and the parameter data frame or matrix (w/o name and type) will be returned.
In the resulting matrix, the item names will be stored as the row names.
\cr
Note that \code{type="B"} means both "B" and "B3".
To select 2PLM items, use \code{type="B2"}. \cr
Also, \code{type="Bn"} means both "Bn" and "Bn3".
To select 2PNM items, use \code{type="Bn2"}. \cr
\cr
If \code{type="ALL"}, regardless of \code{tomat}, the input data frame
will be returned after validity check.
}
\examples{
checkparam(paramA1, type="B" )
checkparam(paramA1, type="B2", tomat=0 )
checkparam(paramA1, type="G" )

}
