% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkparam2.R
\name{checkparam2}
\alias{checkparam2}
\title{Checking and Subsetting the Parameter Data Frame (New Version)}
\usage{
checkparam2(
  param,
  type = NULL,
  modulename = "checkparam2",
  tomat = 1,
  printerror = 1
)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{type}{Item type}

\item{modulename}{name of the module which calls this function}

\item{tomat}{= 0 not to convert the result to a matrix.}

\item{printerror}{= 0 to suppress error messages.}
}
\value{
A parameter matrix or data frame.
}
\description{
Checking and Subsetting the Parameter Data Frame (New Version)
}
\details{
This function checks if the input param is a valid parameter data frame
with name, type, ncat, p1, p2, p3, ...,  variables.\cr
When type = (B | B3 | Bn | Bn3 | G | Gn | P |  PN | N) is given,
the items with the given types will be selected
and the parameter data frame or matrix (w/o name and type) will be returned.
In the resulting matrix, the item names will be stored as the row names.
\cr
Note that \code{type="B"} means both "B" and "B3".
To select 2PLM items, use \code{type="B2"}. \cr
Also, \code{type="Bn"} means both "Bn" and "Bn3".
To select 2PNM items, use \code{type="Bn2"}. \cr
\cr
If \code{type="ALL"}, regardless of \code{tomat}, the input data frame
will be returned after validity check.

Note that, when \code{type} contains more than 2 types, the result will be
sorted according to the order of the appearance of the types in \code{type}.
}
\examples{
checkparam2(paramA1, type="B" )
checkparam2(paramA1, type=c("B","P"), tomat=0 )
checkparam2(paramA1, type=c("Bn2","G"), tomat=0 )

}
