% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv2G.R
\name{conv2G}
\alias{conv2G}
\title{Conversion to Logistic Graded Response Model}
\usage{
conv2G(
  param,
  theta = NULL,
  init = 1,
  paramG = NULL,
  method = 1,
  wtype = 1,
  wmean = 0,
  wsd = 1,
  DinP = 1,
  npoints = 21,
  thmin = -3,
  thmax = 3,
  printGN = 0,
  maxiter = 500,
  eps = 0.000001,
  epsg = 0.000001,
  epsx = 0.000000001,
  print = 1,
  plot = 0
)
}
\arguments{
\item{param}{Item Parameter Data Frame with item types \cr
"B3","Bn","Bn3","P", "Gn"}

\item{theta}{Vector of theta points}

\item{init}{= 1 to use fitG2P, else use equally spaced b-parameters}

\item{paramG}{initial parameter data frame.
This has priority over \code{init}.}

\item{method}{= 1 to use item category response function (icrf)
 to calculate rmse (default) \cr
= 2 to use item category information function (icif) to calculate rmse \cr
= 3 to use item response function (irf) to calculate rmse
 (not yet available), \cr
= 4 to use item information function (iif) to calculate rmse,
 (not yet available), \cr}

\item{wtype}{= 0 not to use dnorm(theta) as the weight}

\item{wmean}{The mean of normal distribution to be used as the weight}

\item{wsd}{The sd of normal distribution to be used as the weight}

\item{DinP}{= 1 to include D=1.7 in logistic function}

\item{npoints}{# of discrete points for theta}

\item{thmin}{Minimum value of discrete theta value}

\item{thmax}{Maximum value of discrete theta value}

\item{printGN}{print level for lazy.mat::GN function}

\item{maxiter}{Maximum # of GN iterations}

\item{eps}{Convergence criterion for the relative improvement of rmse}

\item{epsg}{Convergence crit for the maximum absolute value of the gradient}

\item{epsx}{Convergence crit for the maximum absolute change of
the parameter value}

\item{print}{>= 1 to print result}

\item{plot}{>= 1 to plot result}
}
\value{
A list of: \cr
paramNew: Fitted logistic GRM item parameter data frame\cr
 2PLM or 2PNM items remain unchaged. \cr
paramP: Input GPCM Item Parameter Data Frame\cr
grad: Gradient matrix \cr
wtype, wmean, wsd, method, init \cr
rmse_p: rmse in terms of icrf (minimized with method=1)\cr
rmse_irf: rmse in terms of irf\cr
rmse_ii: rmse in terms of item information\cr
rmse_iic: rmse in terms of item category information
 (minimized with method=2)\cr
icrfNew, icifNew, iifNew, irfNew \cr
icrfOld, icifOld, iifOld, irfOld
}
\description{
Conversion to Logistic Graded Response Model using Weighted Least Squares \cr
Japanese help file: (\link[lazy.irt]{conv2G_JPH})
}
\details{
This function finds the set of Logistic GRM item parameters
which best fit the given icrfs or item info functions of the items
in the input parameter data frame. \cr\cr

If method = 1, this function minimizes\cr
\code{ sum( w*( vec(icrf(theta)) - vec(icrf_GRM(theta|PARAM)) )^2 ) } \cr
with respect to the GRM item parameters, PARAM, \cr
where \code{icrf(theta)} is the icrf of input items, \cr
 \code{icrf_GRM(theta|PARAM)} is the icrf of fitted GRM items, \cr
and \code{w} is the weight vector ( \code{N(wmean,wsd^2)} or 1 ). \cr

If method = 2, this function minimizes\cr
\code{ sum( w*( vec(info_ic(theta)) - vec(info_ic_GRM(theta|PARAM)) )^2 ) }
 \cr
with respect to the GRM item parameters, PARAM, \cr
where \code{info_ic(theta)} is the item category information function
 of input  items and \cr
 \code{info_ic_GRM(theta|PARAM)} is the item category information function
 of the fitted GRM items. \cr

If method = 3, this function minimizes\cr
\code{ sum( w*( info_i(theta) - info_i_GRM(theta|PARAM) )^2 ) } \cr
with respect to the GRM item parameters, PARAM, \cr
where \code{info_i(theta)} is the item information function
 of input items and \cr
 \code{info_i_GRM(theta|PARAM)} is the item information function of
 fitted GRM items. \cr

\cr
When three parameter binary items are included, 2PLM will be fitted.
\cr
Weighted Gauss-Newton method (\code{lazy.mat::GN}) is used for
the minimization with the numerical
Jacobian matrix calculated by \code{lazy.mat::JacobianMat}.
\cr
}
\examples{
# convert 3PLM and GPCM items to logistic GRM items
param <- paramA1[c(2,5,8),]
theta <- seq(-4,4,length=51)
# maxiter for method=2 is too small.
res1 <- conv2G( param, theta, maxiter=20, plot=1, wtype=1, method=1 )
res2 <- conv2G( param, theta, maxiter=2, plot=1, wtype=1, method=2 )

Print(res1$rmse_p, res1$rmse_iic, res1$rmse_ii)
Print(res2$rmse_p, res2$rmse_iic, res2$rmse_ii)


}
