% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convP2P.R
\name{convP2N}
\alias{convP2N}
\title{Convert Partial Credit Item Parameters in Standard Format  (step parameters)
 to Nominal Format}
\usage{
convP2N(param, print = 0, DinP = 1)
}
\arguments{
\item{param}{Item Parameter Data Frame for Partial Credit Items}

\item{print}{= 1 to print result}

\item{DinP}{= 1 to include D=1.7 in logistic function}
}
\description{
Convert Partial Credit Item Parameters in Standard Format  (step parameters)
 to Nominal Format
}
\details{
ICRF of category k of item j at theta, namely, \eqn{p_{jk}(\theta)}\cr
    is defined as\cr
\eqn{p_{jk}(\theta) = Ez_{jk}(\theta)
/ \sum_{k=0}^{ncat[j]-1} Ez_{jk}(theta)}\cr
      where\cr
\cr
   if type = "P" , \cr
    \eqn{Ez_{jk}(\theta) =}
   \eqn{exp( 1.7^{DinP} \sum_{h=0}^k a*_j \; (\theta - b*_{jh}) )} \cr
                                      \eqn{, k=0,1, ..., ncat[j]-1} \cr
   where \eqn{b*_{jk}} is the step parameter with \eqn{b*_{0j}=0}. \cr
\cr
   if type = "PN" and ctype=0, \cr
    \eqn{Ez_{jk}(\theta) = exp( 1.7^{DinP} a_j \; k (\theta - b_{jk}) )} \cr
                                      \eqn{, k=0,1, ..., ncat[j]-1} \cr
   with \eqn{b_{j0}=0}. \cr
\cr
   if type = "PN" and ctype=1, \cr
    \eqn{Ez_{jk}(\theta) = exp(  a_j k \theta + c_{jk}) )} \cr
                                    \eqn{, k=0,1, ..., ncat[j]-1} \cr
   with \eqn{c_{j0}=0}. \cr
\cr
    Note that if \code{DinP=1}, 1.7 will be used. \cr
\cr
    The step parameter, \eqn{b*_{jk}}, is  the value of theta where
     \eqn{P_{jk-1}(\theta)} and \eqn{P_{jk}(\theta)} intersect.\cr
}
