% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weight_df.R
\name{create_weight_df}
\alias{create_weight_df}
\title{Create Weight Data Frame}
\usage{
create_weight_df(param, w = 1, vvec = NULL)
}
\arguments{
\item{param}{Parameter data frame with which the weight data frame be
associated.}

\item{w}{Item weight vector, or a scalar.}

\item{vvec}{Item category weight vector, or NULL}
}
\value{
A list of \cr
weight The weight data frame created. \cr
minmax A vector containing the minimum and the maximum score of the test.\cr
minmax_i A matrix containing the minimum and the maximum score of each item
before applying the item weight \code{w}..
}
\description{
Create Weight Data Frame
}
\details{
Using the information in param, the natural item category weights will be
generated.

The non NA values of the item category weights must be stored in \code{vvec}.
The length of non NA item category weight of item \code{i} is given by
\code{param$ncat[i]}.

\code{w} can be a scalar.
}
\examples{
create_weight_df( paramS1 )
create_weight_df( paramS1, w=c(2,1,2,1) )

create_weight_df( paramS1, vvec=c(0,1, -1,1, 0,1,2,3, -2,2,4,6) )

}
