% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirf_etc.R
\name{dicrf_num}
\alias{dicrf_num}
\title{Numerical Derivatives of ICRF w.r.t. theta}
\usage{
dicrf_num(param, theta, log = 0, second = 0, eps = 0.000001)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{theta}{Discrete theta values}

\item{log}{= 1 to calculate log derivatives.}

\item{second}{= 1 to calculate second derivatives}

\item{eps}{eps for JacobianMat}
}
\value{
dICRF  length(theta) x sum(ncat) matrix
}
\description{
Numerical Derivatives of ICRF w.r.t. theta
}
\details{
This function uses lazy.mat::JacobianMat to calculate the first derivative.
Except for the second derivatives, the analytic functions are available
in \code{dirf}.
}
\examples{
res1 <- dicrf_num( paramS1, -3:3 )
res2 <- dirf( paramS1, -3:3, print=0, plot=0 )$dICRF
Print(apply(abs(res1-res2),2,max))

# second derivative of P
res3 <- dicrf_num( paramS1, -3:3, second=1 )


\dontrun{

thmin <- -4
thmax <- 4
npoints <- 121
theta <- seq(thmin,thmax,length=npoints)

param <- paramA1[1:7,]


log <- 1
res1 <- dicrf_num( param, theta, log=log, second=1 )

# all
matplot(theta,res1, type="l", main="2nd deriv. of logP")

# 0-th category: all negative
fromP <- get_range(param$ncat )[,1]
matplot(theta,res1[,fromP], type="l", main="2nd deriv. of logP: 0-th cat")

# all negative except for the 1st category of 3PLM
matplot(theta,res1[,-c(4,8)], type="l"
        , main="2nd deriv. of logP: all but 1st of 3PLM")

# 1st category of 3PLM
matplot(theta,res1[,c(4,8)], type="l", main="2nd deriv. of logP: 1st of 3PLM")

}

}
