% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirf_etc.R
\name{dicrfBn}
\alias{dicrfBn}
\title{Calculation of the Derivative of Item Response Function of Binary
Normal CDF Items}
\usage{
dicrfBn(param, theta, maxZ = 700, zero = 0, smallP = 0, print = 0, plot = 0)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{theta}{Discrete theta values}

\item{maxZ}{Maximum value of the argument to logistic function}

\item{zero}{= 0 to exclude the xzero-th category from output}

\item{smallP}{Minimum value of probability}

\item{print}{= 1 to print result}

\item{plot}{= 1 to plot result}
}
\value{
dP       length(theta) x nitems   dICRF matrix when zero=0 \cr
           length(theta) x 2*nitems dICRF matrix when zero=1
}
\description{
Calculation of the Derivative of Item Response Function of Binary
Normal CDF Items
}
