% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirf_etc.R
\name{dicrfN}
\alias{dicrfN}
\title{Calculation of the Derivative of Item Response Function of Nominal
 Items}
\usage{
dicrfN(param, theta, smallP = 0, print = 0, plot = 0, debug = 0, DinP = 1)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{theta}{Discrete theta values}

\item{smallP}{Minimum value of probability}

\item{print}{= 1 to print result}

\item{plot}{= 1 to plot result}

\item{debug}{= 1 to print intemediate result}

\item{DinP}{= 1 to include D=1.7 in logistic function}
}
\description{
Calculation of the Derivative of Item Response Function of Nominal
 Items
}
