% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirf_etc.R
\name{dicrfPN0}
\alias{dicrfPN0}
\title{Calculation of the Derivative of Partial Credit ICRF in Nominal format}
\usage{
dicrfPN0(param, theta, smallP = 0, print = 0, plot = 0, debug = 0)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{theta}{Discrete theta values}

\item{smallP}{Minimum value of probability}

\item{print}{= 1 to print result}

\item{plot}{= 1 to plot result}

\item{debug}{= 1 to print intemediate result}
}
\value{
A list of \cr
  dP             length(theta) x sum(ncat) dICRF matrix \cr
  dPt            length(theta) x nitems derivative of
                  item response function \cr
  fromP, toP    indexes of each ICRF in P \cr
  fromPP, toPP  indexes of each ICBRF in PP
}
\description{
Calculation of the Derivative of Partial Credit ICRF in Nominal format
}
