% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirf_etc.R
\name{dirf}
\alias{dirf}
\title{Calculation of Derivative of Item Response Function}
\usage{
dirf(
  param,
  theta = NULL,
  weight = NULL,
  zero = 1,
  smallP = 0.000000001,
  thmin = -4,
  thmax = 4,
  npoints = 21,
  DinP = 1,
  numderiv = 0,
  eps = 0.000001,
  log = 0,
  print = 1,
  debug = 0,
  plot = 0
)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{theta}{Discrete theta values}

\item{weight}{Weight data frame}

\item{zero}{= 0 to exclude the xzero-th category from output}

\item{smallP}{Minimum value of probability}

\item{thmin}{Minimum value of discrete thata value}

\item{thmax}{Maximum value of discrete thata value}

\item{npoints}{# of discrete points for theta}

\item{DinP}{= 1 to include D=1.7 in logistic function}

\item{numderiv}{= 1 to use numerical derivatives}

\item{eps}{eps for JacobianMat}

\item{log}{= 1 to calculate log derivatives}

\item{print}{= 1 to print result}

\item{debug}{= 1 to print intemediate result}

\item{plot}{= 1 to plot result}
}
\value{
A list of \cr
  dICRF, dIRF, dTRF, fromP, toP=toP, vecv, minscore_i, maxscore_i,
  minscore_t, maxscore_tt, log
\cr
  where \cr
    dICRF  npoints x sum(ncat) \cr
    dIRF   npoints x nitems   weighted by item category weight \cr
    dTRF   npoints x 1        weighted by item category weight \cr
                                         and item weight \cr
    fromP, toP    location of each item category in dICRF \cr
    vectorize category weights \cr
    minscore_i mimimum score of each item \cr
    maxscore_i maximum score of each item \cr
    minscore_t mimimum score of test \cr
    maxscore_t maximum score of test
\cr
Note that when log=1, dICRF etc are the log derivatives, namely,
the derivative of log ICRF w.r.t. theta, etc.
}
\description{
Calculation of Derivative of Item Response Function
}
\examples{
dirf( paramS1, plot=1 )

# compare analytic and numeric derivative
res1=dirf( paramS1, print=0, plot=0 )$dICRF
res2=dirf( paramS1, print=0, numderiv=1, plot=0 )$dICRF
Print(max(abs(res1-res2)))

}
