% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirf_etc.R
\name{dirf_p}
\alias{dirf_p}
\title{Calculation of the Derivatives of the Item Category Response Function
with respect to Item Parameters}
\usage{
dirf_p(
  paramj,
  theta = NULL,
  weight = NULL,
  smallP = 0,
  DinP = 1,
  thmin = -4,
  thmax = 4,
  npoints = 21,
  Pj = NULL,
  PPj = NULL,
  zero = 0,
  cat.first = 0,
  log = 0,
  print = 0
)
}
\arguments{
\item{paramj}{item parameters data frame for ONE item}

\item{theta}{Discrete theta values}

\item{weight}{Weight data frame:  NOT used.}

\item{smallP}{Minimum value of probability}

\item{DinP}{= 1 to include D=1.7 in logistic function}

\item{thmin}{Minimum value of discrete thata value}

\item{thmax}{Maximum value of discrete thata value}

\item{npoints}{# of discrete points for theta}

\item{Pj}{icrf:   npoints x (ncatj-1)  (no zero category)  or NULL}

\item{PPj}{icbrf of the Graded Response Model   or NULL}

\item{zero}{= 1 to include the zero-th category in output}

\item{cat.first}{= 1 to chage the category fist in the rows of Jack.}

\item{log}{= 1 to obtain the log Jacobian: d log(ICRF) d param}

\item{print}{= 1 to print result}
}
\value{
list of (Jack, Pj, PPj) \cr
  Jack     (length(theta) x (ncatj-1)) x ncatj \cr
            derivative of vec(Pj) with respect to (a, b1, b2, ...) \cr
          If cat.first = 0 \cr
            theta changes first, then k changes from 1 to ncatj \cr
          If cat.first = 1, category(k) changes first, then theta. \cr
  Pj Item Category Response Function (icrf)  (length(theta) x (ncatj-1))\cr
  PPj will be output when type="G" or "Gn".
}
\description{
Calculation of the Derivatives of the Item Category Response Function
with respect to Item Parameters
}
\examples{
res=dirf_p( paramS1[3,], npoints=5, print=1 )
res=dirf_p( paramS1[3,], npoints=5, print=1, zero=1 )
res=dirf_p( paramS1[3,], npoints=5, print=1, cat.first=1 )


}
