% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_rank.R
\name{est_rank}
\alias{est_rank}
\title{Estimation of the Latent Rank of LRT Model}
\usage{
est_rank(
  Uc = NULL,
  U = NULL,
  V = NULL,
  alpha = rep(1, ncol(V)),
  vmin = 0.0001,
  vmax = 1 - vmin,
  rho = NULL,
  print = 0,
  plot = 0,
  title = NULL
)
}
\arguments{
\item{Uc}{Item Response Data in compressed format}

\item{U}{Item Response Data}

\item{V}{LRT Item Parameter Matrix:   nitem x nclass}

\item{alpha}{= vector of Dirichle parameters of length nclass \cr
or a number A to set alpha=rep(A,nclass) \cr
or a negative number -A to set alpha=A*normal pdf}

\item{vmin}{minimum value of V}

\item{vmax}{maximum value of V}

\item{rho}{initial value of probability vector of each latent class.}

\item{print}{= 1 to print the result}

\item{plot}{= 1 to plot the estimated theta distributions}

\item{title}{Title strings}
}
\value{
A list of: \cr
rank The estimated rank: nrow(Uc) x 1 \cr
H The posterior probability matrix: nrow(Uc) x nclass \cr
rho The prior probability of the class: \cr
alpha The prior of rho: \cr
nclass The number of classes \cr
method EAP
}
\description{
Estimation of the Latent Rank of LRT Model
}
\details{
Note that uLRT returns the V  as the nclass x nitem data frame,
whereas in this function  V is defined as nitem x nclass. (sorry!)
\cr\cr
The core part of uLRT, namely, the E-step, is used in this function.
\cr\cr
The rank of a person i is defined as the location of the highest
posterior probability of H[i,].
\cr
}
\examples{
#### In the following examples, maxiter is set to 20 which is
#### not large enough to obtain convergence.
####
#
#
#
set.seed(1701)

param=paramB1[c(1:3,7:9,13:15),]
thmin=-2; thmax=2; npoint=5
N=1000

# discrete theta
# theta0=seq(thmin,thmax,length=npoint)
theta0=c(-2, -1, 0, 2, 3)
theta=unlist(lapply( theta0, rep, round(N/npoint) ))
res2 <- gendataIRT( 1, paramB1, theta=theta, compress=1 )
Uc=as.data.frame(res2$U)
ncat=res2$ncat
type=res2$type

nclass=5
res1 <- uLRT( Uc, nclass=nclass, estrho=1, monotone=1, alpha=20
, maxiter=20, plot=1, print=1 )
V1=res1$V[,seq(2,2*res1$nitems,2)]
res <- est_rank( Uc=Uc, V=t(V1), print=1, alpha=-4, plot=1, title="Test" )

}
