% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_theta.R
\name{est_theta}
\alias{est_theta}
\title{Estimation of Theta}
\usage{
est_theta(
  Uc = NULL,
  U = NULL,
  param = NULL,
  method = "EAP",
  min_theta = -7,
  max_theta = 7,
  theta = NULL,
  thd = NULL,
  thdist = "NORMAL",
  thmin = -4,
  thmax = 4,
  npoints = 21,
  thmean = 0,
  thstd = 1,
  print = 0,
  plot = 0
)
}
\arguments{
\item{Uc}{The compressed item response data.}

\item{U}{The uncompressed item response data.}

\item{param}{The parameter data frame.}

\item{method}{= "EAP", "MAP" or "ML"}

\item{min_theta}{Minimum value of theta for method="MLE" or "MAP".}

\item{max_theta}{Maximum value of theta for method="MLE" or "MAP".}

\item{theta}{Discrete theta values.}

\item{thd}{pdf of theta normalized to sum to unity. \cr
When NULL, N(0,1)  will be used. \cr
When thd == 1, locally uniform prior will be used.}

\item{thdist}{= "NORMAL" or "UNIFORM"}

\item{thmin}{Minimum value of discrete thata value.}

\item{thmax}{Maximum value of discrete thata value.}

\item{npoints}{# of discrete points for theta.}

\item{thmean}{The prior mean of normal theta distribution.}

\item{thstd}{The prior standard deviation of normal theta distribution.}

\item{print}{= 2 to print the result.}

\item{plot}{= 1 to show the histogram of theta.}
}
\value{
A list of \cr
thetahat Estimated theta values \cr
thetastd Standard Deviation of the Estimated theta values \cr
theta Theta values for EAP \cr
thd Theta disrtibution for EAP\cr
method  \cr
thdist \cr
thmin and thmax \cr
npoints \cr
thmean \cr
thstd \cr
min_theta and max_theta \cr
}
\description{
Estimation of Theta
}
\details{
The core part of EAP is as follows:
\preformatted{
 ULP=U\%*\%t(logP)
 eULP=exp(ULP)*matrix(1,nrow(ULP))\%*\%t(thd)
 rseULP=rowSums(eULP)
 H=eULP/rseULP
 thetahat=H\%*\%theta
}
where \code{logP} is the log of irf matrix (ntheta x sum(ncat)), \cr
\code{U} is the un-compressed item response matrix( N x sum(ncat)), \cr
and \code{thd} is the normalized prior theta pdf vector (ntheta). \cr
Therefore, \code{H} is N x ntheta matrix of normalized posterior pdf \cr
First two lines calculate the unnormalized joint pdf of U and theta. \cr
The next line line normalizes the posterior distribution of theta
given \code{U} stored in \code{H} matrix. \cr
The last line calculates the posterior mean.
\cr\cr
When \code{thd} is a scalar, the locally uniform theta distribution
 is used.
\cr\cr
method = "ML" or method ="MAP" use native optimize function and
may be very slow.

When method is EAP or ML, the posterior standard deviation of theta or
the standard error of theta will be calculated.
}
\examples{

# compressed item response at equally spaced 21 theta points in [-3,3]
set.seed(1701)
resg <- gendataIRT( 1, paramS3, npoints=21, compress=1 )
Uc <- resg$U
theta <- resg$theta
thetaEAP <- est_theta( Uc, param=paramS3, print=1 )$thetahat
thetaMAP <- est_theta( Uc, param=paramS3, print=1, method="MAP" )$thetahat
thetaML <- est_theta( Uc, param=paramS3, print=1, method="ML" )$thetahat
cor( cbind(theta,thetaEAP,thetaMAP,thetaML) )
plot(theta, thetaEAP)

# uncompressed item response at equally spaced 101 theta points in [-3,3]
set.seed(1701)
resg <- gendataIRT( 1, paramS3, npoints=101, compress=0 )
U <- resg$U
theta <- resg$theta
res <- est_theta( U=U, param=paramS3, print=1 )
plot(theta,res$thetahat)

# 100 normally distriuted theta
set.seed(1701)
resg <- gendataIRT( 1, paramS3, npoints=100, compress=1, thdist="rnorm" )
Uc <- resg$U
theta <- resg$theta
res <- est_theta( Uc=Uc, param=paramS3, print=1 )
plot(theta,res$thetahat)

}
