% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_intersection}
\alias{find_intersection}
\title{Find the intersection of icrf}
\usage{
find_intersection(param, print = 0)
}
\arguments{
\item{param}{Item Paramter Data Frame}

\item{print}{= 1 to print the result}
}
\value{
A data frame of nitem x max(ncat)-1 consisting of name, type, ncat and
 the intersections of
0-th and 1st, 1st and 2nd, ..., (ncat-1)-th and ncat-th categories.
}
\description{
Find the intersection of icrf
}
\details{
When item type is "P", the original step parameter b_k
is the intersection of P_{k-1} and P_k, k=1,2,...,ncat-1 \cr
When item type is "G" , the mode of P_k is given as (b_k+b_{k+1})/2 .
\cr
Note that, b_k is stored as p_{k+1} in the param data frame
since p_1 is always the slope.
\cr
The modes of type "P" items and the intersections of type "G" or "B3"
items are found numerically.


\cr\cr
The numerical root finding may result in an error.
}
\examples{
intersection <- find_intersection( paramS1, print=1 )

}
