% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_minmax_score.R
\name{find_minmax_score}
\alias{find_minmax_score}
\title{Find the possible minimum and maximum scores of the test
from the wei ght data frame}
\usage{
find_minmax_score(weight, print = 0)
}
\arguments{
\item{weight}{The weight data frame}

\item{print}{= 1 to print the result}
}
\value{
minmax A vector containing the minimum and the maximum score of the test.\cr
minmax_i A matrix containing the minimum and the maximum score of each item
before applying the item weight \code{w}.
}
\description{
Find the possible minimum and maximum scores of the test
from the wei ght data frame
}
\examples{
find_minmax_score( weightS21, 1 )

}
