% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitG2P2G.R
\name{fitG2P}
\alias{fitG2P}
\title{Approximate Conversion of Partial Credit Items to Graded Response Items
Using Logit Transformation}
\usage{
fitG2P(
  paramP,
  wtype = 1,
  DinP = 1,
  dataframe = 1,
  npoints = 21,
  thmin = -3,
  thmax = 3,
  maxabsv = 1,
  print = 1,
  plot = 0,
  debug = 0
)
}
\arguments{
\item{paramP}{Item Parameter Data Frame for Partial Credit Model}

\item{wtype}{= 1 to use normal weight, else uniform}

\item{DinP}{= 1 to include D=1.7 in logistic function}

\item{dataframe}{= 1 to create parameter data frame, not matrix}

\item{npoints}{# of discrete points for theta}

\item{thmin}{Minimum value of discrete thata value}

\item{thmax}{Maximum value of discrete thata value}

\item{maxabsv}{Maximum absolute value of the weight}

\item{print}{= 1 to print result}

\item{plot}{= 1 to plot result}

\item{debug}{= 1 to print intemediate result}
}
\value{
Graded Response Model Item Parameter Data Frame
}
\description{
Approximate Conversion of Partial Credit Items to Graded Response Items
Using Logit Transformation
}
\examples{
paramG1 <- fitG2P( paramS1[4,], npoints=21, plot=1, print=1 )

}
