% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitI2L.R
\name{fitI2L}
\alias{fitI2L}
\title{Approximate Conversion of LRT to IRT
Using Logit Transformation}
\usage{
fitI2L(V, print = 0, plot = 0, title = NULL)
}
\arguments{
\item{V}{item x class probability matrix}

\item{print}{= 1 to print the estimated IRT item parameters \cr
= 2 to print the irf.}

\item{plot}{= 1 to plot the main result \cr
= 2 to plot irf of each item.}

\item{title}{title string}
}
\value{
A list of: \cr
theta The estimated theta value for each latent rank \cr
param IRT item parameter data.frame \cr
rmse The rmse stat.
}
\description{
Approximate Conversion of LRT to IRT
Using Logit Transformation
}
\details{
The LS criterion in terms of the logit: \cr
\code{ ssq( logit(t(V)) - logit(irf(theta|item parameters) )} \cr
will be minimized by PCA with respect to theta and item parameters.
}
\examples{
#
#### In the following examples, maxiter is set to 20 which is
#### not large enough to obtain convergence.
####
#
#
#
set.seed(1701)

param <- paramB1[c(1:3,7:9,13:15),]
thmin <- -2; thmax <- 2; npoint <- 5
N <- 1000

# discrete theta
# theta0 <- seq(thmin,thmax,length=npoint)
theta0 <- c(-2, -1, 0, 2, 3)
theta <- unlist(lapply( theta0, rep, round(N/npoint) ))
res2 <- gendataIRT( 1, paramB1, theta=theta, compress=1 )
Uc <- as.data.frame(res2$U)
ncat <- res2$ncat
type <- res2$type

# lrt parameters
nclass <- 5
resm1 <- uLRT( Uc, nclass=nclass, estrho=1, monotone=1, alpha=20
                                  , maxiter=20, plot=1, print=1 )
V1 <- resm1$V[,seq(2,2*resm1$nitems,2)]

# conversion
res <- fitI2L( t(V1), print=1, plot=1 )
plot(theta0, res$theta,type="b", main="original theta vs recovered theta")

}
