% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitI2L_ls0.R
\name{fitI2L_ls0}
\alias{fitI2L_ls0}
\title{Conversion of LRT to IRT}
\usage{
fitI2L_ls0(
  V,
  print = 0,
  plot = 0,
  title = NULL,
  wtype = 0,
  maxiter = 1000,
  eps = 0.000001,
  SQUAREM = 3,
  nSQUAREM = 1,
  minalpha = -999,
  maxalpha = -1,
  always = 1,
  reset1 = 0,
  reset2 = 1
)
}
\arguments{
\item{V}{item x class probability matrix}

\item{print}{= 1 to print the estimated IRT item parameters \cr
= 2 to print the irf.}

\item{plot}{= 1 to plot the main result \cr
= 2 to plot irf of each item.}

\item{title}{title string}

\item{wtype}{= 1 to use dnomr(theta) as the weight \cr
= 0 to use no weight.}

\item{maxiter}{Max # of iterations.}

\item{eps}{Convergence criterion for optim.}

\item{SQUAREM}{= 3 :   See the help of iSQUAREM in lazy.accel package.}

\item{nSQUAREM}{when to star iSQUAREM}

\item{minalpha}{= -999 :   See the help of iSQUAREM in lazy.accel package.}

\item{maxalpha}{= -1 :   See the help of iSQUAREM in lazy.accel package.}

\item{always}{= 0 :   See the help of iSQUAREM in lazy.accel package.}

\item{reset1}{= 1 :   See the help of iSQUAREM in lazy.accel package.}

\item{reset2}{= 2 :   See the help of iSQUAREM in lazy.accel package.}
}
\value{
A list of: \cr
theta The estimated theta value for each latent rank \cr
param IRT item parameter data.frame \cr
rmse The rmse stat. \cr
wtype The type of weight.
}
\description{
Conversion of LRT to IRT
}
\details{
The LS criterion: \cr
\code{ ssq( t(V) - irf(theta|item parameters) )} \cr
will be minimized  with respect to theta and item parameters.
}
\examples{
#
#### In the following examples, maxiter is set to 20 which is
#### not large enough to obtain convergence.
####
#
#
#
set.seed(1701)

param <- paramB1[c(1:3,7:9,13:15),]
thmin <- -2; thmax <- 2; npoint <- 5
N <- 1000

# discrete theta
# theta0 <- seq(thmin,thmax,length=npoint)
theta0 <- c(-2, -1, 0, 2, 3)
theta <- unlist(lapply( theta0, rep, round(N/npoint) ))
res2 <- gendataIRT( 1, paramB1, theta=theta, compress=1 )
Uc <- as.data.frame(res2$U)
ncat <- res2$ncat
type <- res2$type

# lrt parameters
nclass <- 5
resm1 <- uLRT( Uc, nclass=nclass, estrho=1, monotone=1, alpha=20
                                  , maxiter=20, plot=1, print=1 )
V1 <- resm1$V[,seq(2,2*resm1$nitems,2)]

# conversion
res <- fitI2L_ls( t(V1), print=1, plot=1, wtype=1, maxiter=50 )
plot(theta0, res$theta,type="b", main="original theta vs recovered theta")

}
