% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitG2P2G.R
\name{fitP2G}
\alias{fitP2G}
\title{Approximate Conversion of Graded Response Items to Partial Credit Items
Using Logit Transformation}
\usage{
fitP2G(
  paramG,
  wtype = 1,
  DinP = 1,
  dataframe = 1,
  npoints = 21,
  thmin = -3,
  thmax = 3,
  maxabsv = 3,
  print = 1,
  plot = 0,
  debug = 0
)
}
\arguments{
\item{paramG}{Item Parameter Data Frame for Graded Response Items}

\item{wtype}{= 1 to use normal weight, else uniform}

\item{DinP}{= 1 to include D=1.7 in logistic function}

\item{dataframe}{= 1 to create parameter data frame, not matrix}

\item{npoints}{# of discrete points for theta}

\item{thmin}{Minimum value of discrete thata value}

\item{thmax}{Maximum value of discrete thata value}

\item{maxabsv}{Maximum absolute value of the weight}

\item{print}{= 1 to print result}

\item{plot}{= 1 to plot result}

\item{debug}{= 1 to print intemediate result}
}
\value{
A list of:
Partial Credit Model Item Parameter Data Frame in Nominal Format \cr
Partial Credit Model Item Parameter Data Frame in Nominal Format0 \cr
Partial Credit Model Item Parameter Data Frame in Standard Format
}
\description{
Approximate Conversion of Graded Response Items to Partial Credit Items
Using Logit Transformation
}
\examples{
paramP1 <- fitP2G( paramS1[3,], npoints=21, plot=1, print=1 )$paramP
paramG1 <- fitG2P( paramP1, npoints=21, plot=1, print=1 )



}
