% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_SEM_theta.R
\name{flatten_SEM_theta}
\alias{flatten_SEM_theta}
\title{Find a Transformation g of the Thetahat based Observed Score X such that
Y=g(X) has a Flat Standard Error of Measurement.}
\usage{
flatten_SEM_theta(
  out_obscore = NULL,
  sigma = 1,
  by_s = 0.1,
  param = NULL,
  weight = NULL,
  npoints = 131,
  thmin = -4,
  thmax = 4,
  thdist = 1,
  alpha = 0.1,
  compress = 0,
  print = 1,
  plot = 0,
  debug = 0
)
}
\arguments{
\item{out_obscore}{The result of obscore function \cr
This has the priority over the set of the arguments of obscore function.}

\item{sigma}{The standard error of the transformed score}

\item{by_s}{The interval for continuous S.}

\item{param}{Item Parameter Data Frame for obscore}

\item{weight}{Weight data frame for obscore}

\item{npoints}{# of discrete points for theta for obscore}

\item{thmin}{Minimum value of discrete thata value for obscore}

\item{thmax}{Maximum value of discrete thata value for obscore}

\item{thdist}{Type of theta distribution for obscore \cr
= 0 to use uniform,  = 1 to use N(0,1)}

\item{alpha}{small prob for quantile and confidence interval for obscore}

\item{compress}{= 1 to remove zero-probability weighted total observed
scores for obscore}

\item{print}{> 1 to print result}

\item{plot}{> 1 to plot result}

\item{debug}{= 1 to print intemediate result}
}
\value{
A list of the following: \cr
theta: The value of theta \cr
stdx_theta: SEM of X at theta \cr
s: The transformed true score: Y=g(X) and s=g(t) \cr
gdash: The derivative of g \cr
stdy_s: SEM of Y at s \cr
lengtht: length of t \cr
sigma: New SEM value specified \cr
brk_x2u: Break points of X to create S.\cr
brk_x2uc: Break points of X to create almost condinuous S.\cr\cr
out_obscore: The output from the obscore function.
}
\description{
Find a Transformation g of the Thetahat based Observed Score X such that
Y=g(X) has a Flat Standard Error of Measurement.
}
\details{
Let \code{stdx(t)} be the standard error of measurement of X at t. \cr
This can be calculated as \code{stdx}_t by the obscore function. \cr
The standard deviation of \code{Y=g(X)} at t can be approximated by \cr
 \code{ g-dash(t)*stdx(t) }\cr
and we want it to be a constant (sigma). \cr
Therefore, \cr
 \code{ g-dash(t) = sigma / stdx(t)) } \cr
and the \code{g} function can be recovered by integrating the above
\code{g-dash}. \cr
This \code{g} is the vaiance-stabilizing transformation.

Notes: \cr
 Recommended to use \code{npoints=151, thmin=-4, thmax=4} or larger
 for obscore function.
}
\examples{

# tiny set of binary items
param=paramB1
maxscore=sum(param$ncat-1)
param$p1=1
out_obscore <- obscore( param )
res=flatten_SEM_theta( out_obscore, sigma=1, plot=1, print=1 )

# binary and polytomous items
res2=flatten_SEM_theta( param=paramS1, sigma=1, plot=1, print=1 )


}
