% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_icrfnames.R
\name{gen_icrfnames}
\alias{gen_icrfnames}
\title{Generate the Row Names of vec(icrf) as the Combination of Theta Values
and Category Names.}
\usage{
gen_icrfnames(
  theta,
  ncat,
  zero = 0,
  cat.first = 0,
  digits = 2,
  catroot = "c",
  sep = ":",
  compress = 0
)
}
\arguments{
\item{theta}{Numeric vector of theta values.}

\item{ncat}{# of categories}

\item{zero}{= 1 to include the 0-th category}

\item{cat.first}{= 1 to chage the category fist. \cr
= 2 to change the category first and place category name before theta. \cr
This is for vec(icrf, byrow=1 ).}

\item{digits}{# of digits after the decimal point of theta.}

\item{catroot}{The root character for category.}

\item{sep}{The caracter separating theta and category.}

\item{compress}{= 1 to remove all the blanks in the result}
}
\value{
A vector consisting of rownames of vec(icrf) or vec(icrf, byrow=1).
}
\description{
Generate the Row Names of vec(icrf) as the Combination of Theta Values
and Category Names.
}
\details{
This function creates a character vector consisting of \cr
  \code{ c( thetavalue1:catvalue1,  thetavalue2:catvalue1, ... ) }. \cr
If cat.first=1, the result is \cr
 \code{ c( thetavalue1:catvalue1,  thetavalue1:catvalue2, ... ) }, \cr
and if cat.first=2, \cr
 \code{ c( catvalue1:thetavalue1,  catvalue2:thetavalue1, ... ) }, \cr
}
\examples{
gen_icrfnames(seq(-4,4,length=7), 3 )
gen_icrfnames(seq(-4,4,length=7), 3, cat.first=1 )

}
