% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfuncs.R, R/gendataIRT.R
\name{gendataIRT}
\alias{gendataIRT}
\title{Generation of Simulated Item Response Data}
\usage{
gendataIRT(
  Ntot,
  param,
  DinP = 1,
  Nmat = NULL,
  zero = 1,
  theta = NULL,
  npoints = 31,
  thmin = -3,
  thmax = 3,
  thd = NULL,
  thdist = "NORMAL",
  thmean = 0,
  thstd = 1,
  nomiss = 0,
  compress = 0,
  sort = 1,
  thetaname = NULL
)

gendataIRT(
  Ntot,
  param,
  DinP = 1,
  Nmat = NULL,
  zero = 1,
  theta = NULL,
  npoints = 31,
  thmin = -3,
  thmax = 3,
  thd = NULL,
  thdist = "NORMAL",
  thmean = 0,
  thstd = 1,
  nomiss = 0,
  compress = 0,
  sort = 1,
  thetaname = NULL
)
}
\arguments{
\item{Ntot}{# of total observations to be generated
If Ntot == 1, one observation per each theta point
will be generated, \cr
otherwize, the # of observations per each theta point is \cr
proportional to round(Ntot*thd) where thd is \cr
the npoint x 1 vector of theta distribution. \cr
Therefore, Ntot must be large enough. \cr}

\item{param}{Item Parameter Data Frame}

\item{DinP}{= 1 to include D=1.7 in logistic function}

\item{Nmat}{npoints x nitems matrix consisting of # of trials, or NULL.
This will be used only when Ntot=1 and compress=0.}

\item{zero}{= 0 to exclude zero-th category when not compressed.}

\item{theta}{Discrete theta values}

\item{npoints}{# of discrete points for theta, or # of random variables}

\item{thmin}{Minimum value of discrete thata value}

\item{thmax}{Maximum value of discrete thata value}

\item{thd}{theta distribution probability vector}

\item{thdist}{= "NORMAL" or "UNIFORM" or "rnorm"  or  "runif" \cr
When "NORMAL" or "rnorm", \cr
    thmean and thstd will be used to generate thd. \cr
When "UNIFORM" or "runif", \cr
    thmin and thmax will be used to generate thd. \cr
When "rnorm" or "runif", theta will be generated using \cr
    npoints random numbers and thd  and Ntot is set equal to 1.}

\item{thmean}{mean of theta distribution}

\item{thstd}{std of theta distribution}

\item{nomiss}{= 1 to exclude those theta points with N = 0.}

\item{compress}{= 1 to compress the data when Ntot = 1.}

\item{sort}{= 1 to sort the result according to theta.}

\item{thetaname}{= Name of the variable which will contain theta value
or NULL.}
}
\value{
A list of \cr
 U, N, npoints, theta, thd, Ntot, fromP, toP, type, ncat, thmean, thstd
 , compress, sort

\cr
where
\cr
\cr
           If compress = 0 and Ntot is large, \cr
             U    is   npoints x sum of ncat[j] \cr
              U[,fromP[j]:toP[j]]  is npoints x ncat[j] \cr
              U[,fromP[j]:toP[j]][i,k] contains # of responses to \cr
              the k-th category (k=0,1,...,ncat[j]) at theta[i] \cr
              to which N[i]=round(Ntot*thd[i]) observations belong. \cr
\cr
\cr
           If compress = 1 and Ntot = 1, \cr
             U    is   npoints x nitems \cr
              U[i,j] contains the response to the j-th item at theta[i] \cr
              where 0 <= U[i,j] <= ncat[j]-1. \cr
             In this case, fromP and toP are not compressed. \cr
\cr
           (theta, thd, N)   are   npoints x 1 \cr
           N[i]  is the # of observations at theta[i] \cr

A list of \cr
 U, N, npoints, theta, thd, Ntot, fromP, toP, type, ncat, thmean, thstd
 , compress, sort

\cr
where
\cr
\cr
           If compress = 0 and Ntot is large, \cr
             U    is   npoints x sum of ncat[j] \cr
              U[,fromP[j]:toP[j]]  is npoints x ncat[j] \cr
              U[,fromP[j]:toP[j]][i,k] contains # of responses to \cr
              the k-th category (k=0,1,...,ncat[j]) at theta[i] \cr
              to which N[i]=round(Ntot*thd[i]) observations belong. \cr
\cr
\cr
           If compress = 1 and Ntot = 1, \cr
             U    is   npoints x nitems \cr
              U[i,j] contains the response to the j-th item at theta[i] \cr
              where 0 <= U[i,j] <= ncat[j]-1. \cr
             In this case, fromP and toP are not compressed. \cr
\cr
           (theta, thd, N)   are   npoints x 1 \cr
           N[i]  is the # of observations at theta[i] \cr
}
\description{
Generation of Simulated Item Response Data

Generation of Simulated Item Response Data
}
\examples{
# 10 observations with theta from N( 0, 1 ): not compressed
set.seed(1702)
res1 <- gendataIRT( 1, paramS1, npoints=10, thdist="rnorm" )
Print(res1$N,res1$U,res1$theta)
#
# 10 observations with theta from N( 0, 1 ): compressed
set.seed(1702)
res1 <- gendataIRT( 1, paramS1, npoints=10, thdist="rnorm", compress=1 )
Print(res1$N,res1$U,res1$theta)
#
# 10 observations with theta from N( 0, 1 ): no zero category
set.seed(1702)
res1 <- gendataIRT( 1, paramS1, npoints=10, thdist="rnorm", zero=0 )
Print(res1$N,res1$U,res1$theta)
#
#
#
# seven theta points in [-3,3]
set.seed(1701)
res1 <- gendataIRT( 1, paramS1, npoints=7, thmin=-3, thmax=3 )
Print(res1$N,res1$U,res1$theta)
#
# seven theta points in [-3,3] and compressed
set.seed(1701)
res1 <- gendataIRT( 1, paramS1, npoints=7, thmin=-3, thmax=3, compress=1 )
Print(res1$N,res1$U,res1$theta)
#
set.seed(1701)
# 100 observations tabulated at seven theta points in [-3,3]
res1 <- gendataIRT( 1000, paramS1, npoints=7, thmin=-3, thmax=3 )
Print(res1$N,res1$U,res1$theta)
#
#
# Using Nmat argument
set.seed(1701)
npoints=20
param=paramS1
nitems=nrow(param)
Nmat=matrix(100,npoints,nitems)
res1 <- gendataIRT( 1, param, Nmat=Nmat, npoints=npoints
                  , thdist="rnorm", compress=1 )
Print(res1$N,res1$U,res1$theta)


# 10 observations with theta from N( 0, 1 ): not compressed
set.seed(1702)
res1 <- gendataIRT( 1, paramS1, npoints=10, thdist="rnorm" )
Print(res1$N,res1$U,res1$theta)
#
# 10 observations with theta from N( 0, 1 ): compressed
set.seed(1702)
res1 <- gendataIRT( 1, paramS1, npoints=10, thdist="rnorm", compress=1 )
Print(res1$N,res1$U,res1$theta)
#
# 10 observations with theta from N( 0, 1 ): no zero category
set.seed(1702)
res1 <- gendataIRT( 1, paramS1, npoints=10, thdist="rnorm", zero=0 )
Print(res1$N,res1$U,res1$theta)
#
#
#
# seven theta points in [-3,3]
set.seed(1701)
res1 <- gendataIRT( 1, paramS1, npoints=7, thmin=-3, thmax=3 )
Print(res1$N,res1$U,res1$theta)
#
# seven theta points in [-3,3] and compressed
set.seed(1701)
res1 <- gendataIRT( 1, paramS1, npoints=7, thmin=-3, thmax=3, compress=1 )
Print(res1$N,res1$U,res1$theta)
#
set.seed(1701)
# 100 observations tabulated at seven theta points in [-3,3]
res1 <- gendataIRT( 1000, paramS1, npoints=7, thmin=-3, thmax=3 )
Print(res1$N,res1$U,res1$theta)
#
#
# Using Nmat argument
set.seed(1701)
npoints=20
param=paramS1
nitems=nrow(param)
Nmat=matrix(100,npoints,nitems)
res1 <- gendataIRT( 1, param, Nmat=Nmat, npoints=npoints
                  , thdist="rnorm", compress=1 )
Print(res1$N,res1$U,res1$theta)


}
