% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOptWeight.R
\name{GOptWeight}
\alias{GOptWeight}
\title{Estimation of the Globally Optimal Item Category Weights}
\usage{
GOptWeight(
  param,
  i.weight = NULL,
  c.weight = NULL,
  npoints = 31,
  thmin = -4,
  thmax = 4,
  th.m = 0,
  th.sd = 1,
  method = "nlm",
  L = 100,
  eps = 0.000001,
  print = 0
)
}
\arguments{
\item{param}{Item paramter data frame}

\item{i.weight}{A vector of item weights: w}

\item{c.weight}{A vector of item category weights: v}

\item{npoints}{# of discrete theta points}

\item{thmin}{Minimum value of theta points}

\item{thmax}{Maximum value of theta points}

\item{th.m}{Mean of theta dist}

\item{th.sd}{Standard deviation of theta dist}

\item{method}{= "nlm"}

\item{L}{max # of iterations}

\item{eps}{eps for gradient}

\item{print}{= 1 to print result}
}
\description{
This program estimates a set of globally optimal item category weights
which maximizes the expected test imformation function.
}
\examples{
res1 <- GOptWeight( paramS1 )
create_weight_df( paramS1, vvec=res1 )

res2 <- GOptWeight( paramS2 )
create_weight_df( paramS2, vvec=res2 )

}
\author{
Contributed by Dr. Sayaka Arai of DNC.
}
