% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graded_prob.R
\name{graded_prob}
\alias{graded_prob}
\title{Calculate the probability contents associated with the integer score}
\usage{
graded_prob(
  x,
  mean,
  std,
  min = NULL,
  max = NULL,
  truncate = 0,
  method = 1,
  print = 0,
  plot = 0
)
}
\arguments{
\item{x}{The integer score vector}

\item{mean}{The mena of the underlying continuous variable}

\item{std}{The standard deviationi of the underlying continuous variable}

\item{min}{The minimum value of x}

\item{max}{The maximum value of x}

\item{truncate}{= 1 to truncate the range of X between [min,max]}

\item{method}{= 1 to use pnorm(x[i+1])-pnorm(x[i]) \cr
= 0 to set the probabilities proportional to dnorm(x)}

\item{print}{= 1 to print the result}

\item{plot}{= 1 to plot the result}
}
\value{
p A vector (same size as x)
}
\description{
Calculate the probability contents associated with the integer score
}
\details{
Let Y be distributed as N( mean, std ).
This function evaluates the probabilities:
  p[1] = Pr( min < X <= x[1] ), p[2] =  Pr( x[1] < X <= x[2] ),  ...
  p[length(x)] =  Pr( x[n] < X <= max ) \cr\cr

If method = 1, \cr
 p[i] = pnorm( b[i+1], mean, std ) - pnorm( b[i], mean, std )  \cr
else \cr
 p[i] = dnorm( x[i], mean, std ) \cr

If method = 2 or method = 1 and truncate = 1, \cr
the probabilites will be normalized so that  sum(p) = 1.
\cr
When truncate = 1, if the interval (min, max) is not wide enough to
cover the whole range of X, the resulting probabilities may not
sum to unity.
}
\examples{
# The rightmost/leftmost categoris are wider than the rest.
# truncation has no effect
graded_prob( 0:4,  2, 2,  truncate=0, print=1  )
graded_prob( 0:4,  2, 2,  truncate=1, print=1  )
graded_prob( 0:4,  2, 2,  method=2, print=1  )

# The rightmost/leftmost categoris have the same length as the rest.
graded_prob( 0:4,  2, 2,  min=-0.5, max=4.5, truncate=0  )
graded_prob( 0:4,  2, 2,  min=-0.5, max=4.5, truncate=1  )

}
