% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfuncs.R
\name{icrfGn}
\alias{icrfGn}
\title{Calculation of Item Response Function of Graded Response Items
 with Normal Ogive Model (2PNM)}
\usage{
icrfGn(param, theta, smallP = 0, print = 0, plot = 0, debug = 0)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{theta}{Discrete theta values}

\item{smallP}{Minimum value of probability}

\item{print}{= 1 to print result}

\item{plot}{= 1 to plot result}

\item{debug}{= 1 to print intemediate result}
}
\value{
list \cr
  P             length(theta) x sum(ncat) ICRF matrix \cr
  PP            length(theta) x sum(ncat) ICBRF matrix \cr
 Pt            length(theta) x nitems item response function \cr
  fromP, toP    indexes of each ICRF in P \cr
  fromPP, toPP  indexes of each ICBRF in PP \cr
}
\description{
Calculation of Item Response Function of Graded Response Items
 with Normal Ogive Model (2PNM)
}
\details{
An Example of Item Parameter datastet:  maxncat = 5\cr
    Note that the subscript of the b-parameter ranges between\cr
         1 and ncat-1 (= maxscore)  (We set c0=0.)\cr
\cr
    name  ncat     A            b1     b2     b3     b4\cr
\cr
    Item1  2     1.0           0.0     .      .      .\cr
    Item2  2     1.0           0.0     .      .      .\cr
    Item3  3     1.5          -1.0    0.0     .      .\cr
    Item4  4     1.5          -1.0    0.0    1.0     .\cr
    Item5  5     1.5          -1.0    0.0    1.0    2.0\cr
}
