% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfuncs.R
\name{icrfP}
\alias{icrfP}
\title{Calculation of Item Response Function of Partial Credit Items}
\usage{
icrfP(paramP, theta, DinP = 1, smallP = 0, print = 0, plot = 0, debug = 0)
}
\arguments{
\item{paramP}{Item Parameter Data Frame of Partial Credit Items}

\item{theta}{Discrete theta values}

\item{DinP}{= 1 to include D=1.7 in logistic function}

\item{smallP}{Minimum value of probability}

\item{print}{= 1 to print result}

\item{plot}{= 1 to plot result}

\item{debug}{= 1 to print intemediate result}
}
\value{
A list of: \cr
  P       length(theta) x sum(ncat) icrf matrix \cr
  PP      length(theta) x sum(ncat) icbf  matrix \cr
  Pt      length(theta) x nitems item response function \cr
}
\description{
Calculation of Item Response Function of Partial Credit Items
}
\details{
An Example of Item Parameter datastet:  maxncat = 5\cr
    Note that the subscript of the b-parameter ranges between\cr
         1 and ncat-1 (= maxscore)  (We set c0=0.)\cr
\cr
          ncat     A            b1     b2     b3     b4\cr
\cr
    Item3  3     1.5          -1.0    0.0     .      .\cr
    Item4  4     1.5          -1.0    0.0    1.0     .\cr
    Item5  5     1.5          -1.0    0.0    1.0    2.0\cr
\cr
    ICRF of category k of item j at theta, namely, \eqn{p_{jk}(theta)}\cr
    is defined as\cr
     \eqn{p_{jk}(theta) = Ez_{jk}(theta)
     / sum_{k=0}^{ncat[j]-1} Ez_{jk}(theta)}\cr
      where\cr
 \eqn{Ez_{jk}(theta) = exp( 1.7^DinP a*_j k (theta - sum_{h=0}^k b*_{jh}) )}
     \cr
              \eqn{= exp( 1.7^DinP a*_j k (theta - sum_{h=1}^k b*_{jh}) )}
                    \cr
                                     \eqn{, k=0,1, ..., ncat[j]-1}\cr
    with \eqn{b*_{j0}=0}.\cr
\cr
    This \eqn{b*_{jk}} is the original step parameter and it is\cr
    the value of theta where \eqn{P_{jk-1}} and \eqn{P_{jk}} intersect.\cr
}
