% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfuncs.R
\name{icrfPN0}
\alias{icrfPN0}
\title{Calculation of  Partial Credit ICRF in Nominal format}
\usage{
icrfPN0(param, theta, smallP = 0, print = 0, plot = 0, debug = 0)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{theta}{Discrete theta values}

\item{smallP}{Minimum value of probability}

\item{print}{= 1 to print result}

\item{plot}{= 1 to plot result}

\item{debug}{= 1 to print intemediate result}
}
\value{
A list of \cr
  P             length(theta) x sum(ncat) ICRF matrix \cr
  PP            length(theta) x sum(ncat) ICBRF matrix \cr
  Pt            length(theta) x nitems item response function \cr
  fromP, toP    indexes of each ICRF in P \cr
  fromPP, toPP  indexes of each ICBRF in PP
}
\description{
Calculation of  Partial Credit ICRF in Nominal format
}
\details{
An Example of Item Parameter datastet:  maxncat = 5\cr
    Note that the subscript of the c-parameter ranges between\cr
         1 and ncat-1 (= maxscore)  (We set c0=0.)\cr
\cr
          ncat     A            c1     c2     c3     c4\cr
\cr
    Item1  2     1.0           0.0     .      .      .\cr
    Item2  2     1.0           0.0     .      .      .\cr
    Item3  3     1.5          -1.0    0.0     .      .\cr
    Item4  4     1.5          -1.0    0.0    1.0     .\cr
    Item5  5     1.5          -1.0    0.0    1.0    2.0\cr
\cr
    ICRF of category k of item j at theta, namely, \eqn{p_{jk}(theta)}\cr
    is defined as\cr
     \eqn{p_{jk}(theta) = Ez_{jk}(theta)
     / sum_{k=0}^{ncat[j]-1} Ez_{jk}(theta)}\cr
      where\cr
     \eqn{Ez_{jk}(theta) = exp( z_{jk}(theta) )}\cr
       and\cr
     \eqn{z_{jk}(theta) = a_j k theta + c_{k}, k=0,1, ..., ncat[j]-1}\cr
    with \eqn{c_{j0}=0}.\cr
}
