% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iif.R
\name{iif}
\alias{iif}
\title{Calculation of Information Functions}
\usage{
iif(
  param,
  theta = NULL,
  npoints = 31,
  thmin = -4,
  thmax = 4,
  legend = 1,
  maxinfo = 0,
  numderiv = 0,
  smallP = 0.000000001,
  print = 1,
  plot = 0,
  debug = 0
)
}
\arguments{
\item{param}{Item Parameter Data Frame.}

\item{theta}{Discrete theta values}

\item{npoints}{# of discrete points for theta.}

\item{thmin}{Minimum value of discrete thata value.}

\item{thmax}{Maximum value of discrete thata value.}

\item{legend}{= 0 to skip printing legend.}

\item{maxinfo}{= The maximum value of item information function for plot.}

\item{numderiv}{= 1 to use numerical first derivatives of irf.}

\item{smallP}{Minimum value of probability in irf and dirf functions.}

\item{print}{= 1 to print the summary \cr
= 2 to print test information functions. \cr
= 3 to print item information functions. \cr
= 4 to print item category information functions.}

\item{plot}{= 1 to plot test information functions\cr
= 2 to plot item information functions. \cr
= 3 to plot the locally best item and item categorie weights.}

\item{debug}{= 1 to print intemediate result.}
}
\value{
A list of \cr\cr
 theta:    theta points\cr
 fromP, toP: location of each item category in item info \cr
 TRF: test response function (tcc) \cr
 icrf: item category response functions\cr
 dicrf: derivative of icrf \cr
 info:   test information function (if)\cr
 info_item:   item information function (iif) \cr
 info_item_cat:   item category information function (icif)
}
\description{
This function calculates Test (tif), Item (iif) and Item Category (icif)
Information Functions. \cr
Japanese help file: (\link[lazy.irt]{iif_JPH})
}
\details{
The item category information function, icif,
 (item response information function)
is defined as \cr
\eqn{ I_{kj}(\theta) = (P'_{kj}(\theta))^2 / P_{kj}(\theta) - }
 \eqn{ P''_{kj}(\theta) } \cr
where \eqn{P_{kj}(\theta)} is the item category response function and 
\eqn{P'_{kj}(\theta)} is the first derivative of \eqn{P_{kj}(\theta)}.
 \cr
The second derivatives, \eqn{ P''_{kj}(\theta) }, will be calculated
numerically by \code{dirt_num} using \code{lazy.mat::JacobianMat}.

The item information function, iif, is defined as \cr
\eqn{  I_j(\theta) = \sum_k I_{kj}(\theta) }
  \eqn{= \sum_k (P'_{kj}(\theta))^2 / P_{kj}(\theta) } \cr
The test information function, if, is the sum of the above:\cr
 \eqn{  I(\theta) = \sum_j I_j(\theta) }
 \eqn{ = \sum_j \sum_k (P'_{kj}(\theta))^2 / P_{kj}(\theta) }.

Note that
\eqn{  \sum_k P_{kj}(\theta) = 1 } and \eqn{ \sum_k P'_{kj}(\theta) =}
\eqn{ \sum_k P''_{kj}(\theta) = 0 } .

Above corresponds to the information functions from \code{info_func}
or \code{obscore} associated with locally best item category weights (LO).
}
\examples{
resInfo <- iif( paramA1, plot=4, print=4 )

}
\references{
Birnbaum, A.(1968) Some Latent Traint Models.
In F. M. Lord and M. R. Novick, Statistical Theories of Mental Test Scores.
 Reading, Mass.: Addison-Wesley. \cr

Samejima,  F.  (2010). The General Graded Response Model. (p79-80)
In Nering, M. L. and Ostini, R. Eds. Handbook of Polytomous
Item Response Theory Models. NY, NY: Routledge

Samejima,  F.  (1969). Estimation  of  a  latent  ability  using  a
response  pattern  of  graded  scores.  (Eq 6-6 in p39)
Psychometrika  Monographs, 34
(Suppl. 4).
}
