% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iif_JPH.R
\docType{package}
\name{iif_JPH}
\alias{iif_JPH}
\title{情報関数の計算}
\usage{
iif( param, theta=NULL, npoints=31, thmin=-4, thmax=4
, legend=1, maxinfo=0, numderiv=0, smallP=1e-9
, print=1, plot=0, debug=0 )
}
\arguments{
\item{param}{項目パラメタデータフレーム}

\item{theta}{特性値 theta の離散点}

\item{smallP}{確率の最小値}

\item{thmin}{特性値 theta の最小値}

\item{thmax}{特性値 theta の最大値}

\item{npoints}{離散点の数}

\item{numderiv}{= 1 irf の１次の数値微分を用いる。}

\item{print}{= 1 要約の出力\cr
= 2 テスト情報関数 (tif) の出力. \cr
= 3 項目情報関数 (iif) の出力. \cr
= 4 項目カテゴリ情報関数 (icif) の出力.}

\item{plot}{= 1 tif のプロット\cr
= 2 iif のプロット. \cr
= 3 局所的最適重みと項目カテゴリ重みのプロット.}

\item{legend}{= 0 レジェンドの印刷をしない。}

\item{maxinfo}{= プロットにおける情報関数の最大値.}

\item{debug}{= 1 中間結果を出力する}
}
\value{
以下の要素からなるリスト
 \preformatted{
 theta: 特性値 theta の値
 TRF: テスト反応関数 (trf)
 icrf: 項目カテゴリ反応関数 (icrf)
 dicrf: icrf の１次微分
 info: テスト情報関数 (tif)\cr
 info_item: 項目情報関数 (iif) \cr
 info_item_cat:項目カテゴリ情報関数 (icif)
 fromP, toP:  icrf や info_item_cat の列中の各項目（のカテゴリ）の位置
 }
}
\description{
この関数は、テスト情報関数 (tif), 項目情報関数 (iif) ならびに
項目カテゴリ情報関数 (icif) を計算する \cr
English help file: (\link[lazy.irt]{iif})
}
\details{
項目カテゴリ情報関数 icif は以下のように定義されている。 \cr
\eqn{ I_{kj}(\theta) = (P'_{kj}(\theta))^2 / P_{kj}(\theta) - }
 \eqn{ P''_{kj}(\theta) } \cr
ただし、icrf を \eqn{P_{kj}(\theta)} とすれば、
\eqn{P'_{kj}(\theta)} は icrf の１次微分、\eqn{P'_{kj}(\theta)} は
２次微分である。
なお、
\eqn{ P''_{kj}(\theta) } は \code{dicrf_num} 関数において
 \code{lazy.mat::JacobianMat} を用いて数値的に計算されている。.

項目情報関数 (iif) とテスト情報関数 (tif) は以下のように定義されている。 \cr
\eqn{  I_j(\theta) = \sum_k I_{kj}(\theta) }
  \eqn{= \sum_k (P'_{kj}(\theta))^2 / P_{kj}(\theta) } \cr
 \eqn{  I(\theta) = \sum_j I_j(\theta) }
 \eqn{ = \sum_j \sum_k (P'_{kj}(\theta))^2 / P_{kj}(\theta) }.


なお、これらの情報関数は \code{info_func} や \code{obscore} が出力する
局所的最適重み LO を用いた情報関数と一致している。.
}
\examples{
resInfo <- iif( paramA1, plot=4, print=4 )

# @keywords internal
}
\references{
Birnbaum, A.(1968) Some Latent Traint Models.
In F. M. Lord and M. R. Novick, Statistical Theories of Mental Test Scores.
 Reading, Mass.: Addison-Wesley.

前川眞一 (2023) 項目反応理論におけるモデル変換. 日本テスト学会誌 19(1) 35-58.


Samejima,  F.  (2010). The General Graded Response Model. (p79-80)
In Nering, M. L. and Ostini, R. Eds. Handbook of Polytomous
Item Response Theory Models. NY, NY: Routledge

Samejima,  F.  (1969). Estimation  of  a  latent  ability  using  a
response  pattern  of  graded  scores.  (Eq 6-6 in p39)
Psychometrika  Monographs, 34
(Suppl. 4).
}
