% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_func.R
\name{info_func}
\alias{info_func}
\title{Calculation of Various Information Functions}
\usage{
info_func(
  param,
  theta = NULL,
  weight = NULL,
  npoints = 31,
  thmin = -4,
  thmax = 4,
  thdist = "NORMAL",
  thmean = 0,
  thstd = 1,
  legend = 1,
  maxinfo = 0,
  numderiv = 0,
  smallP = 0.000000001,
  print = 1,
  plot = 0,
  debug = 0
)
}
\arguments{
\item{param}{Item Parameter Data Frame.}

\item{theta}{Discrete theta values}

\item{weight}{Weight data frame.}

\item{npoints}{# of discrete points for theta.}

\item{thmin}{Minimum value of discrete thata value.}

\item{thmax}{Maximum value of discrete thata value.}

\item{thdist}{= "NORMAL" or "UNIFORM" \cr
When "NORMAL" thmean and thstd will be used to generate thd. \cr
When "UNIFORM" thmin and thmax will be used to generate thd. \cr}

\item{thmean}{mean of theta distribution}

\item{thstd}{std of theta distribution}

\item{legend}{= 0 to skip printing legend.}

\item{maxinfo}{= The maximum value of item information function for plot.}

\item{numderiv}{= 1 to use numerical derivatives.}

\item{smallP}{Minimum value of probability in irf and dirf functions.}

\item{print}{= 1 to print the summary \cr
= 2 to print test information functions. \cr
= 3 to print item information functions. \cr
= 4 to print the locally best item categorie weights.}

\item{plot}{= 1 to plot test information functions\cr
= 2 to plot item information functions. \cr
= 3 to plot the locally best item and item categorie weights.}

\item{debug}{= 1 to print intemediate result.}
}
\value{
A list of \cr\cr
 theta:    theta points\cr
 thd: theta distribution (to calculate average info) \cr
 info : information function defind as \code{(slope_TRF)^2 / (stdx_t)^2}\cr
 fromP, toP: location of each item category in item info \cr
 TRF: test response function (tcc) \cr
 TRF_LO: test response function (tcc) with locally best weight\cr
 icrf: item category response functions\cr
 dicrf: derivative of icrf \cr
 info_LOW:   information function with locally best item weight\cr
 info_LO:   information function with locally best category weight\cr
 info_item_LOW:  item information function
             with locally best item weight\cr
 info_item_LO:   item information function
             with locally best category weight \cr
 stdx_t: SEM at theta (conditional standard deviation of X given theta) \cr

 w_LO: the locally best item weight \cr
 v_LO: the locally best item category weight
 scaled so that v[0,j]=0, j=1,2,...,nitems. \cr
 A: The Basic Function \cr\cr
 a_stdx_t: square root of the average of stdx_t^2 \cr
 a_info: average of info \cr
 a_info_LOW: average of info_LOW \cr
 a_info_LO: average of info_LO \cr
 a_info_item_LOW: average of info_item_LOW (for each item)\cr
 a_info_item_LO: average of info_item_LO (for each item) \cr
 a_w_LO: average of w_LO (for each item) \cr
 a_v_LO: average of v_LO (for each item category) \cr
}
\description{
This function calculates the information functions associated with
the set of given weights, \cr
and two types of locally optimam weights.
}
\details{
Note that, given category and item weights, information function is defined
as \cr
 \code{ ( slope of TRF at theta )^2 / (variance of x at theta) }\cr
where TRF and x is calculated with the given set of weights. \cr
\cr
In general. the best weights depends on the value of theta.
Therefore, the name  locally best weight. \cr
The best item weight given category weights is called as
locally best item weight, or LOW. \cr
When the category weights themselves are optimized it is called as
the locally best weight, or, LO. \cr
This weight is equivalent to the basic function of Samejima(1969) .\cr
\cr
The information function with LO is defined as \cr
\eqn{ \sum_j \sum_k (P'_{kj}(\theta))^2 / P_{kj}(\theta) } \cr
where \eqn{P_{kj}(\theta)} is the item category response function, \cr
and \eqn{P'_{kj}(\theta)} is its derivative. \cr
\cr
The information function with LOW is defined as \cr
\eqn{ \sum_j  (P_j^{*'}(\theta))^2 / var(U_j^{*} | \theta) } \cr
where \eqn{U_j^{*} = \sum_k v_{kj} U_{kj}} is
 the weighted item score,\cr
and \eqn{P_j^{*'}(\theta)} is the derivative of the expected value of
\eqn{U_j^{*}} at theta. \cr
\cr\cr
The Basic Function is defined as \eqn{ P`(theta) / P(theta) }. \cr
The Best Item Category Weight = Basic Function - Basic Function
of the 0-th category. \cr
\cr
Item Info Function is defined as
  \eqn{ \sum_k (P'_{kj}(\theta))^2 / P_{kj}(\theta) }.
}
\examples{
resInfo <- info_func( paramS2, plot=2, print=1 )
resInfo <- info_func( paramS2, weight=weightS21, plot=1, print=0 )
resInfo <- info_func( paramS2, weight=weightS22, plot=1, print=0 )

}
\references{
Birnbaum, A.(1968) Some Latent Traint Models.
In F. M. Lord and M. R. Novick, Statistical Theories of Mental Test Scores.
 Reading, Mass.: Addison-Wesley. \cr

Samejima,  F.  (1969). Estimation  of  a  latent  ability  using  a
response  pattern  of  graded  scores. Psychometrika  Monographs, 34
(Suppl. 4).
}
