% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invtrf.R
\name{invtrf}
\alias{invtrf}
\title{Inverse Function of trf}
\usage{
invtrf(
  param,
  x = NULL,
  xmin = NULL,
  xmax = NULL,
  by_x = 1,
  weight = NULL,
  theta = NULL,
  thmin = -5,
  thmax = 5,
  npoints = 121,
  method = 0,
  interpol = "spline",
  print = 0,
  plot = 0
)
}
\arguments{
\item{param}{Item parameter data frame}

\item{x}{The value of x, the observed test score, or NULL}

\item{xmin}{The minimum value of x, or NULL}

\item{xmax}{The maximum value of x, or NULL}

\item{by_x}{The increment of x values}

\item{weight}{The weight data frame or the vector of item weight, or NULL}

\item{theta}{The value of theta for method=0, or NULL}

\item{thmin}{The minimum value of theta for method=0, or NULL}

\item{thmax}{The maximum value of theta for method=0, or NULL}

\item{npoints}{The # of theta points for method=0}

\item{method}{= 1 to use native \code{uniroot} function (slow) \cr
= 0 to use interpolation with option \code{lazy.tools::interpol} (fast).}

\item{interpol}{The interpolation method. See \code{lazy.tools::interpol}.}

\item{print}{= 2 to print the result}

\item{plot}{= 1 to plot the result}
}
\value{
A list of \cr
\code{x} The value of x.\cr
\code{theta} The value of theta. \cr
\code{locminmax} A list containing the locations of -9999 or 9999.
}
\description{
Inverse Function of trf
}
\details{
The test response function, trf, is calculated as:
\preformatted{
 trf(theta) = sum_{j=1}^n w_j x sum_{k=0}^{ncat[i]-1} v_{kj} P_{kj}(theta)
 where
 w_j is the item weight for item j, (default=1)
 v_{kj} is the item category weight for the k-th category of item j.
 The default value of v_{kj} = k, k=0, 1, 2, ... ncat[i]-1
}
When values \code{x=NULL} the value of \code{x} will be calculated
using the weight information.

If x is too small or too large, -9999 or 9999 will be returned.
}
\examples{
res <- invtrf( paramS1, method=1, print=2 )
res <- invtrf( paramS1, method=0, print=2 )
res <- invtrf( paramS1, weight=weightS12, method=1, print=2 )
res <- invtrf( paramS1, weight=weightS12, method=0, print=2 )

res <- invtrf( paramS1, by_x=0.5, print=2 )
res <- invtrf( paramS1, weight=c(1,1,2,2), print=2 )

}
