% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfuncs.R
\name{irf}
\alias{irf}
\title{Calculation of Item Response Function}
\usage{
irf(
  param,
  theta = NULL,
  weight = NULL,
  zero = 1,
  smallP = 0.000000001,
  thmin = -4,
  thmax = 4,
  npoints = 21,
  DinP = 1,
  print = 1,
  debug = 0,
  plot = 0,
  colors = "black",
  plotchar = NULL,
  linetype = NULL
)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{theta}{Discrete theta values}

\item{weight}{Weight data frame}

\item{zero}{= 0 to exclude the zero-th category from output}

\item{smallP}{Minimum value of probability}

\item{thmin}{Minimum value of discrete thata valuepackage}

\item{thmax}{Maximum value of discrete thata value}

\item{npoints}{# of discrete points for theta}

\item{DinP}{= 1 to include D=1.7 in logistic function}

\item{print}{= 1 to print result}

\item{debug}{= 1 to print intemediate result}

\item{plot}{= 1 to plot result}

\item{colors}{= color of the lines in plot}

\item{plotchar}{= pch vector to be used in the plot}

\item{linetype}{= lty vector to be used in the plot}
}
\value{
list( ICRF, IRF, TRF, fromP, toP )\cr
\cr
   where\cr
     ICRF  npoints x sum(ncat)\cr
     IRF   npoints x nitems   weighted by item category weight\cr
     TRF   npoints x 1        weighted by item category weight\cr
                                          and item weight\cr
     fromP, toP    location of each item category in ICRF\cr
\cr
 Needs:\cr
   icrfB, icrfBN, icrfG, idrfGn, icrfPN, icrfP, icrfN, checkparam\cr
\cr
}
\description{
This function calculates icrf, irf and trf. \cr
Japanese help file: (\link[lazy.irt]{irf_JPH})
}
\details{
It is assumed that the data frame has the following columns
  in the following order.\cr
\cr
  name, type, ncat, p1, p2, ..., p[ncat+1]\cr
\cr
\cr
  When item_type == "B"  or "B2"  or  "B3"   or   "Bn", ( ncat == 2 )\cr
    p1 is the discrimination parameter\cr
    p2 is the difficulty parameter\cr
    p3 is the gussing parameter\cr
\cr\cr
  When item_type == "G" or "Gn",\cr
    p1 is the discrimination parameter\cr
    p2 is the category threshold parameter b_{j1}\cr
    p3 is the category threshold parameter b_{j2}\cr
\cr
    p_ncat is the category threshold parameter b_{j,[ncat-1]}\cr
\cr\cr
  When item_type == "P",\cr
    p1 is the discrimination parameter\cr
    p2 is the step parameter b_{j1}\cr
    p3 is the step parameter b_{j2}\cr
\cr
    p_ncat is the step parameter b_{j,[ncat-1]}\cr
\cr\cr
  When item_type == "PN,\cr
    p1 is the slope parameter\cr
    p2 is the intercept parameter b_{j1}\cr
    p3 is the intercept parameter b_{j2}\cr
\cr
    p_ncat is the intercept parameter b_{j,[ncat-1]}\cr
\cr\cr
  When item_type == "N,\cr
    p1 - p_(ncat[j]-1) are the slope parameters\cr
    p_ncat[j] - p_2*(ncat[j]-1) are the intercept parameters\cr
\cr
  Regardless of the item_type, there will be ncat[j]\cr
  item paramters for item j, except for the Binary Items\cr
  which have the gusseing param as the ncat[j]+1st,\cr
  and the Nominal Items which have 2*(ncat[j]-1) item parameters.\cr
}
\examples{
Cards="
name type ncat   p1   p2   p3   p4
  Q1    B    2  0.9  0.0,,
  Q2    B3   2  1.0  0.0  0.2,
  Q3    G    4  1.0 -2.0  0.0  2.0
  Q4    P    4  0.8 -2.0  0.0  2.0
"; param <- cards( Cards, header=1 )
irf( param, plot=1 )

}
