% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obscore.R
\name{obscore}
\alias{obscore}
\title{Calculation of Observed Score Distribution and Posterior Distribution
of Theta with Various Information Functions}
\usage{
obscore(
  param,
  weight = NULL,
  npoints = 31,
  thmin = -4,
  thmax = 4,
  thdist = 1,
  alpha = 0.1,
  compress = 0,
  print = 1,
  plot = 0,
  debug = 0
)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{weight}{Weight data frame \cr
or a vector containing item weights. \cr
or a scalar or NULL.}

\item{npoints}{# of discrete points for theta}

\item{thmin}{Minimum value of discrete thata value}

\item{thmax}{Maximum value of discrete thata value}

\item{thdist}{Type of theta distribution \cr
= 0 to use uniform,  \cr = 1 to use N(0,1)}

\item{alpha}{small prob for quantile and confidence interval}

\item{compress}{= 1 to remove zero-probability weighted total observed
scores}

\item{print}{= 1 to print result}

\item{plot}{= 1 to plot result}

\item{debug}{= 1 to print intemediate result}
}
\value{
\cr
  list( theta_stat, obs_stat, Px_t, Pt_x,  etcetc )\cr
\cr
   where\cr
\cr
  theta_stat as data frame containing the following:\cr
    theta:       theta points\cr
    Pt:          prior probability distribution of theta\cr
    TRF:         test response function\cr
    slope_TRF:   slope of TRF\cr
    stdx_t:      standard deviation of X (observed score) given theta\cr
    info:        information function defined as
    (slope_TRF)^2 / (stdx_t)^2\cr
    info_LOW:    information function with the locally optimal item weight
    given category weights\cr
    info_LO:     information function with the locally optimal category
     weight\cr
    qt_L:        upper quantile of X given theta\cr
    qt_U:        lower quantile of X given theta\cr
    poststd:     posterior std of theta given X\cr
                   as a function of posterior mean\cr
\cr
  obs_stat as data frame containing the following:\cr
    score:      domain of X (observed score)\cr
    Px:         marginal probability  of X\cr
    post_mean:  posterior mean of theta given X\cr
    post_std:   posterior standard deviation of theta given X\cr
    ci_L:       upper limit of conficence interval for theta given X\cr
    ci_U:       lower limit of conficence interval for theta given X\cr
    ci_hwid:    half the width of CI\cr
\cr
  SigmaX2, SigmaT2, SigmaE2: (observed, true, and average error variances)
\cr
  aSEM=sqrt(SigmaE2): average Standard Error of Measurement, \cr
  rel: test reliability
\cr
  SigmaT_theta2, SigmaE_theta2, aSEM_theta, rel_theta
\cr
  Px_t           score x theta  conditional prob of X given theta\cr
  Pt_x           score x theta  conditional prob of theta given X\cr
  Px             score x 1      marginal dist of X\cr
\cr
pdfname: item parameter data frame name \cr
wdfname: item weight data frame name \cr
npoints: # of theta points \cr
thmin, thmax: the range of theta \cr
thdist: Type of theta distribution \cr
nitems: # of items \cr
minscore_t: minimum score \cr
maxscore_t: maximum score \cr
alpha: small probability value \cr
}
\description{
Japanese help file: (\link[lazy.irt]{obscore_JPH})
}
\details{
This function calculates the distribution of the weighted observed score,
the information function associated with it, and the posterior
distribution of theta given the observed score. \cr
Also, the information functions associated with two types of
locally optimum weights will be calculated.

Note that, given category and item weights, information function is defined
as \cr
\code{
 ( slope of TRF at theta )^2 / (variance of X at theta)
 }\cr
where TRF and X are calculated with the given set of weights. \cr
This is the information function associated with the test scores \cr
calculated using the given item and  category weights. \cr
\cr
The information can be increased by adjusting the weights: \cr
However, in general, the optimal weights which maximize the information
function depend on the value of theta. \cr
Therefore, the name  "locally" optimal weight. \cr\cr
The optimal item weight given category weights is called as
the locally optimal item weight, or LOW and defined as:. \cr
\eqn{ w_j(\theta)= P^{*'}_j(\theta) / var( U_j^{*} | \theta) } \cr
where
\eqn{ U_j^{*} = \sum_k v_{kj} U_{kj}} and
\eqn{ P^{*'}_j(\theta)=\sum_k v_{kj} P_{kj}(\theta) }.
\cr
When the category weights themselves are optimized it is called as
the locally optimal weights, or, LO, \cr
which are related to the basic function of Samejima(1969)
defined as:.\cr
\eqn{ v_{kj}(\theta)=P'_{kj}(\theta) / P_{kj}(\theta)
 - P'_{0j}(\theta) / P_{0j}(\theta) }.
\cr\cr
The information function with LO is defined as \cr
\eqn{ \sum_j \sum_k (P'_{kj}(\theta))^2 / P_{kj}(\theta) } \cr
where \eqn{P_{kj}(\theta)} is the item category response function, \cr
and \eqn{P'_{kj}(\theta)} is its derivative. \cr
This is the maximum information function for all the theta range.
\cr
The information function with LOW is defined as \cr
\eqn{ \sum_j  (P_j^{*'}(\theta))^2 / var(U_j^{*} | \theta) } \cr
where \eqn{U_j^{*} = \sum_k v_{kj} U_{kj}} is
 the weighted item score,\cr
and \eqn{P_j^{*'}(\theta)} is the derivative of the expected value of
\eqn{U_j^{*}} at theta. \cr

\cr
The test reliability coefficient is calculated as
\code{rel = (SigmaX^2-SigmaE^2)/SigmaX^2}, \cr
where SigmaE^2 is the square root of the average of stdx_t^2, \cr
SigmaX^2 is the variance of the observed score.
\cr
}
\examples{
# Define the observed raw score X
# and calculate the score distribution, information function, etc.
res <- obscore( paramS1, plot=1 )

# Define X using the item weights w stored in weightS11
res <- obscore( paramS1, weight=weightS11, plot=1 )

# Define X using the item and category weights w and v stored in weightS12.
res <- obscore( paramS1, weight=weightS12, plot=1 )

}
\references{
Birnbaum, A.(1968) Some Latent Traint Models.
In F. M. Lord and M. R. Novick, Statistical Theories of Mental Test Scores.
 Reading, Mass.: Addison-Wesley. \cr

Mayekawa, S. (2018) A Method to Estimate the Ability from the Weighted
Total Score wich Minimizes the Standard Error of Estimation.
DNC Research Note. RN-18-01. \cr

Mayekawa, S., & Arai, S. (2008).
Distribution of the Sum of Scored Multinomial Random Variables
and Its Application to the Item Response Theory.
In K. Shigemasu, A. Okada, T.Imaizumi, & T. Hoshino (Eds.)
New Trends in Psychometrics. Tokyo: University Academic Press. \cr

Samejima,  F.  (1969). Estimation  of  a  latent  ability  using  a
response  pattern  of  graded  scores. Psychometrika  Monographs, 34
(Suppl. 4).
}
