% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obscore_s.R
\name{obscore_s}
\alias{obscore_s}
\title{Calculate Observed Score Distribution (simple version of \code{obscore})}
\usage{
obscore_s(
  param,
  weight,
  thmin = -4,
  thmax = 4,
  npoints = 31,
  thdist = 1,
  print = 0
)
}
\arguments{
\item{param}{Parameter data frame}

\item{weight}{Weight data frame}

\item{thmin}{Minimum value of discrete thata value}

\item{thmax}{Maximum value of discrete thata value}

\item{npoints}{# of discrete points for theta}

\item{thdist}{Type of theta distribution \cr
= 0 to use uniform,  \cr = 1 to use N(0,1)}

\item{print}{= 1 to print the result}
}
\description{
Calculate Observed Score Distribution (simple version of \code{obscore})
}
\details{
This function calculates the marginal distribution of the test score. \cr
\preformatted{
 1. Calculate the conditional distribution of X given theta
 2. Calculate the joint distribution of X and theta.
 3. Calcukate the marginal distribution of X.
}
In step 1 above, \code{icrf} is calculated by \code{lazy.irt::irf}
and the distribution of the weighted sum of scored multinomial distributions
will be calculated by \code{lazy.irt::sumsmnw}.
}
\examples{
res <- obscore_s( paramS2, weightS21, print=1 )

}
