% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_reg.R
\name{ordinal_reg}
\alias{ordinal_reg}
\title{Oridinal Regression}
\usage{
ordinal_reg(
  R,
  x,
  type = "P",
  param = NULL,
  maxiter = 100,
  eps = 0.00000001,
  epsd = 0.00001,
  smallP = 0,
  DinP = 1,
  minp1 = 0.01,
  maxabsparam = 20,
  print = 1
)
}
\arguments{
\item{R}{A matrix consisting of frequency of criterion variable:
lenght(x) x ncat \cr
R[,1] = Frequency of the zero-th category at x,
R[,2] = Frequency of the first category at x, up to R[,ncat]}

\item{x}{A vector or matrix of unidimensional regressor variable}

\item{type}{Item type  "B", "B3", "G", "P", "PN", "N"}

\item{param}{A data frame containing the initial value of parameters,
or  NULL}

\item{maxiter}{Max # of iterations}

\item{eps}{Criterion for convergence in llh}

\item{epsd}{Criterion for convergence in absolute value of parameter value}

\item{smallP}{Smallest value of probability}

\item{DinP}{= 0 to omit 1.7}

\item{minp1}{Min value of the slope parameter for type != "N" items.}

\item{maxabsparam}{Max value of parameters}

\item{print}{= 0 to surpress output}
}
\value{
A list of: \cr
param Parameter data frame \cr
llh Maximized log likelihood \cr
P Probability at x \cr
x input \cr
g gradient \cr
R input \cr
maxag Maximum of absolute gradient values
}
\description{
This program regresses R on x where
R is the length(x) x ncat frequency matrix and x is the univariate
regressor vector. \cr
Japanese help file: (\link[lazy.irt]{ordinal_reg_JPH})
}
\details{
Let R be n x ncat matrix, and x, n x 1.  \cr
This function maximizes the following log likelihood
w.r.t. item parameters \eqn{param}. \cr
\eqn{llh = \sum_{i=1}^{n} \sum_{k=1}^{ncat}R[i,k] log( P_k(x[i]|param) }
\cr
where \eqn{P_k(x[i]|param)} is the item response function defined by
\code{type} argument and \eqn{param} is its item parameters.
\cr
When \eqn{P_k} is 3PLM, \eqn{c} parameters will be transformed to
\eqn{y=logit(c)} and \eqn{llh} will be maximized w.r.t. \eqn{y}. \cr
Note that \cr
\eqn{d llh / d y = d llh / d c \times c(1-c)}, \cr
where \eqn{d logistic(y) / d y = c(1-c)}
}
\examples{
# binary data
set.seed(1701)
theta=seq(-4,4,length.out=51)
resg=gendataIRT( 500, paramS1[1,], theta=theta, thdist="NORMAL", thd=NULL
               , thmean=0, thstd=1, compress=0 )
testdata=as.matrix( resg$U )

# 3PLM
res=ordinal_reg( testdata, theta, type="B3", param=NULL, print=1 )
# 2PLM
res=ordinal_reg( testdata, theta, type="B", param=NULL, print=1 )
res=ordinal_reg( testdata, theta, type="P", param=NULL, print=1 )
# 2PLM with nonzero c parameter
initp=paramS1[2,]; initp$type="B"
res=ordinal_reg( testdata, theta, type="B", param=initp, print=1 )


# polytomous data
set.seed(1701)
theta=seq(-4,4,length.out=51)
resg=gendataIRT( 500, paramS1[3,], theta=theta, thdist="NORMAL", thd=NULL
                 , thmean=0, thstd=1, compress=0 )
testdata=as.matrix( resg$U )

# graded response model
res=ordinal_reg( testdata, theta, type="G", param=NULL, print=1 )
# partial credit model
res=ordinal_reg( testdata, theta, type="P", param=NULL, print=1 )
res=ordinal_reg( testdata, theta, type="PN", param=NULL, print=1 )


}
