% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oseq.R
\name{oseq}
\alias{oseq}
\title{IRT Observed Score Equating}
\usage{
oseq(
  param1 = NULL,
  param2 = NULL,
  weight1 = NULL,
  weight2 = NULL,
  thmin = -4,
  thmax = 4,
  npoints = 31,
  thdist = 1,
  smooth1 = 0,
  bandwid1 = 3,
  smooth2 = 0,
  bandwid2 = 3,
  round = 0,
  print = 0,
  plot = 0
)
}
\arguments{
\item{param1}{The item parameter data frame for test1}

\item{param2}{The item parameter data frame for test2}

\item{weight1}{The weight data frame for test1, or NULL}

\item{weight2}{The weigth data frame for test2, or NULL}

\item{thmin}{The minimum value of theta to calculate
the marginal score distribution.}

\item{thmax}{The maximum value of theta to calculate
the marginal score distribution.}

\item{npoints}{# of theta points in \code{[thmin,thmax]}}

\item{thdist}{= 1 to use normal theta, = 0 to use uniform theta.}

\item{smooth1}{# of times to smooth cdf1.}

\item{bandwid1}{bandwidth for running average smooth of cdf1.}

\item{smooth2}{# of times to smooth cdf2.}

\item{bandwid2}{bandwidth for running average smooth of cdf2.}

\item{round}{= 1 to round the result to integer}

\item{print}{= 1 or 2 to print the result}

\item{plot}{= 1 to plot the result}
}
\value{
A list of \cr
x1 The test score of test1. \cr
x2 The test score of test2. \cr
x2_1 The test2 equivalent score of x1. \cr
p1 The marginal distribution of test1 score. \cr
p2 The marginal distribution of test2 score.
}
\description{
This function performs the IRT Observed Score Equating of
two test scores. \cr
Japanese help file: (\link[lazy.irt]{oseq_JPH})
}
\details{
This function equates the test score of test1, x1, to the score of test2,
x2. by the IRT Observed Score Equating.\cr
The resulting score will be named as x2_1.

In irt observed score equating, the conditional distribution of X
given theta, Px_t, is calculated first.
Then, using the marginal distribution of theta, P_t,
the joint distribution of X and theta, Pxt, is calculated.
Finally, the marginal distribution of X, Px, is calculated from Pxt. \cr
Given (x1, Px1)  and (x2, Px2), equipercentile equating will be used
to equate x1 to x2.


Note that, when \code{round=1}, the result may NOT be symmetric.\cr
Run the example below with \code{by_x=1} and \code{round=1}.

The marginal distribution of the test score will be calculated by
\link[lazy.irt]{obscore_s} which is a simpler form of 
\link[lazy.irt]{obscore}.

The equipercentile equating will be performed by \link[lazy.irt]{coseq}.
}
\examples{
res=oseq( paramS1, paramS2, print=2 )
res=oseq( paramS1, paramS2, weight1=weightS12, weight2=weightS21, print=2 )

# The effect of item weight.
res=oseq( paramS1, paramS1, weight2=weightS12, print=3 )

# checking if symmetric
# equate test1 to test2
res1 <- oseq( paramS1, paramS2 )
# equate test2 to test1
res2 <- oseq( paramS2, paramS1 )
# merge result
r1 <- data.frame(x1=res1$x1, y1=res1$x2_1)
r2 <- data.frame(y2=res2$x1, x2=res2$x2_1)
rx <- merge( r1, r2, by.x="x1", by.y="x2", all=TRUE)
rx$y <- ifelse( is.na(rx$y1), rx$y2, rx$y1 )
ry <- merge( r1, r2, by.x="y1", by.y="y2", all=TRUE)
ry$x <- ifelse( is.na(ry$x1), ry$x2, ry$x1 )
rxr <- round(rx,4); ryr=round(ry,4)
printm(rxr,ryr)

}
