% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{paramA1}
\alias{paramA1}
\title{Item Parameter Data Frame # 1. (10 items of all types)\cr}
\format{
A data frame with 7 rows and 6 variables.
\describe{
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{p1}{num: First item parameter: slope: a}
 \item{p2}{num: 2nd  item parameter: threshold: b1}
 \item{p3}{num: 3rd  item parameter: threshold: b2 or c-parameter}
 \item{p4}{num: 4th  item parameter: threshold: b3}
}
}
\source{
Generated by author.
}
\usage{
paramA1
}
\description{
Item Parameter Data Frame # 1. (10 items of all types)\cr
}
\details{
\preformatted{
name type ncat   p1      p2      p3      p4
  Q1    B    2  1.0     0.0    0.0
  Q2    B3   2  1.0     0.0    0.2
  Q3    Bn   2  1.0     0.0    0.0
  Q4    Bn3  2  1.0     0.0    0.2
  Q5    P    3  1.0    -1.0    1.0
  Q6    G    3  1.0664 -1.0288 1.0288
  Q7    Gn   3  1.0819 -1.0221 1.0221
  Q8    P    4 0.8000 -2.0000 0.0000 2.0000
  Q9    G    4 0.9179 -2.0495 0.0000 2.0495
 Q10   Gn    4 0.9396 -2.0245 0.0000 2.0245
  }
   item type = all types \cr
   ncat is the max # of categories \cr
 \cr\cr
 The type P item was converted to type G and Gn items using
 fitG2P_ls and fitGn2P_ls functions, resp, so that
 the icrfs of type G and Gn items are as close as
 the irf of type P item.
}
\keyword{datasets}
