% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{paramB1}
\alias{paramB1}
\title{Binary Item Parameter Data Frame # 1. (18 items)\cr}
\format{
A data frame with 18 rows and 6 variables
\describe{
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{p1}{num: First item parameter: slope: a}
 \item{p2}{num: 2nd  item parameter: threshold: b}
 \item{p3}{num: 3rd item parameter: lower asymptote: c}
}
}
\source{
Generated by author.
}
\usage{
paramB1
}
\description{
Binary Item Parameter Data Frame # 1. (18 items)\cr
}
\details{
\preformatted{
name type ncat   p1   p2   p3
  Q1    B    2  1.0 -1.0  0.0
  Q2    B    2  1.0  0.0  0.0
  Q3    B    2  1.0  1.0  0.0
  Q4    B    2  1.0 -1.0  0.0
  Q5    B    2  1.0  0.0  0.0
  Q6    B    2  1.0  1.0  0.0
  Q7    B    2  1.3 -1.0  0.0
  Q8    B    2  1.3  0.0  0.0
  Q9    B    2  1.3  1.0  0.0
  Q10   B    2  1.3 -1.0  0.0
  Q11   B    2  1.3  0.0  0.0
  Q12   B    2  1.3  1.0  0.0
  Q13   B    2  0.7 -1.0  0.0
  Q14   B    2  0.7  0.0  0.0
  Q15   B    2  0.7  1.0  0.0
  Q16   B    2  0.7 -1.0  0.0
  Q17   B    2  0.7  0.0  0.0
  Q18   B    2  0.7  1.0  0.0
  }
   item type = "B" \cr
   ncat is the max # of categories \cr
}
\keyword{datasets}
