% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{paramB2}
\alias{paramB2}
\title{Binary Item Parameter Data Frame # 2. (18 items)\cr}
\format{
A data frame with 18 rows and 6 variables
\describe{
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{p1}{num: First item parameter: slope: a}
 \item{p2}{num: 2nd  item parameter: threshold: b}
 \item{p3}{num: 3rd item parameter: lower asymptote: c}
}
}
\source{
Generated by author.
}
\usage{
paramB2
}
\description{
Binary Item Parameter Data Frame # 2. (18 items)\cr
}
\details{
\preformatted{
name type ncat   p1   p2   p3
  Q1    B3    2  1.0 -1.0  0.2
  Q2    B3    2  1.0  0.0  0.2
  Q3    B3    2  1.0  1.0  0.2
  Q4    B3    2  1.0 -1.0  0.2
  Q5    B3    2  1.0  0.0  0.2
  Q6    B3    2  1.0  1.0  0.2
  Q7    B3    2  1.3 -1.0  0.2
  Q8    B3    2  1.3  0.0  0.2
  Q9    B3    2  1.3  1.0  0.2
  Q10   B3    2  1.3 -1.0  0.2
  Q11   B3    2  1.3  0.0  0.2
  Q12   B3    2  1.3  1.0  0.2
  Q13   B3    2  0.7 -1.0  0.2
  Q14   B3    2  0.7  0.0  0.2
  Q15   B3    2  0.7  1.0  0.2
  Q16   B3    2  0.7 -1.0  0.2
  Q17   B3    2  0.7  0.0  0.2
  Q18   B3    2  0.7  1.0  0.2
  }
   item type = "B3"  (3PLM)\cr
   ncat is the max # of categories \cr
}
\keyword{datasets}
