% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{paramCal1}
\alias{paramCal1}
\title{Small Item Parameter Data Frame for Testing cala and calr #1.\cr}
\format{
A data frame with 33 rows and 8 variables. \cr
Basically, the structure is the same as the parameter data frame \cr
except that the first column containing the form name is added.\cr
\describe{
 \item{gfid}{character:  Form name}
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{p1}{num: First item parameter: slope}
 \item{p2}{num: 2nd  item parameter}
 \item{p3}{num: 3rd item parameter}
}
}
\source{
Generated by author.
}
\usage{
paramCal1
}
\description{
This data frame contains three sets of estimated item parameters of
binary items distributed across three test forms. \cr
The item parameters were estimated independently for each form.
}
\details{
\preformatted{
    gfid name type ncat        p1          p2        p3
form1  QB1   B2    2 1.0000000 -1.00000000 0.0000000
form1  QB2   B2    2 1.0000000  0.00000000 0.3000000
form1  QB4   B3    2 0.8000000  1.00000000 0.2000000
form1  QB5   B2    2 1.0000000 -1.30000000 0.0000000
form1  QB6   B2    2 1.0000000  0.30000000 0.3000000
form1  QB7   B3    2 1.0000000  1.30000000 0.2000000
form1  QB8   B3    2 0.8000000  0.70000000 0.2000000
form1 QB11   B2    2 1.2000000 -1.00000000 0.0000000
form1 QB12   B2    2 1.2000000  0.00000000 0.3000000
form1 QB14   B3    2 1.0000000  1.00000000 0.2000000
form1 QB15   B2    2 0.8000000 -1.00000000 0.0000000
form1 QB16   B2    2 0.8000000  0.00000000 0.3000000
form1 QB17   B3    2 0.8000000  1.00000000 0.2000000
form1 QB18   B3    2 0.7000000  1.00000000 0.2000000
form2  QB1   B2    2 0.9736860 -2.57272092 0.0000000
form2  QB3   B3    2 0.7560111  0.04149805 0.3905005
form2  QB5   B2    2 0.7700730 -2.90974572 0.0000000
form2  QB7   B3    2 1.0288158  0.65955944 0.2237003
form2 QB11   B2    2 1.1783210 -2.51973224 0.0000000
form2 QB13   B3    2 1.0667175  0.20900095 0.4426091
form2 QB15   B2    2 0.7748544 -2.41498863 0.0000000
form2 QB17   B3    2 0.6003595  0.18576845 0.2764597
form2  QB2   B2    2 0.9845137 -1.11132899 0.3000000
form2  QB6   B2    2 0.9434385 -0.75960242 0.3000000
form3  QB2   B2    2 1.4515893  0.91079758 0.3000000
form3  QB4   B3    2 1.2179065  1.75344958 0.2100962
form3  QB6   B2    2 1.5139847  1.16847211 0.3000000
form3  QB8   B3    2 1.1706377  1.56182045 0.3502745
form3 QB12   B2    2 1.6937542  0.99698111 0.3000000
form3 QB14   B3    2 1.4473557  1.75645583 0.3786156
form3 QB16   B2    2 1.1834637  0.79398007 0.3000000
form3 QB18   B3    2 0.9699614  1.79807714 0.4093079
form3  QB1   B2    2 1.4712936  0.17616872 0.0000000
}
}
\keyword{datasets}
