% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{paramCal2}
\alias{paramCal2}
\title{Small Item Parameter Data Frame for Testing cala and calr #2.\cr}
\format{
A data frame with 40 rows and 8 variables. \cr
Basically, the structure is the same as the parameter data frame \cr
except that the first column containing the form name is added.\cr
\describe{
 \item{gfid}{character:  Form name}
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{p1}{num: First item parameter: slope}
 \item{p2}{num: 2nd  item parameter}
 \item{p3}{num: 3rd item parameter}
 \item{p4}{num: 4yth item parameter}
}
}
\source{
Generated by author.
}
\usage{
paramCal2
}
\description{
This data frame contains three sets of estimated item parameters of
various types of items distributed across three test forms. \cr
The item parameters were estimated independently for each form.
}
\details{
\preformatted{
 gfid name type ncat        p1          p2         p3       p4
form1  QB1   B2    2 1.0000000 -1.00000000  0.0000000       NA
form1  QB2   B2    2 1.0000000  0.00000000  0.3000000       NA
form1  QB4   B3    2 0.8000000  1.00000000  0.2000000       NA
form1  QG1    G    3 1.0000000 -2.00000000  0.0000000       NA
form1  QG2    G    3 1.0000000 -1.00000000  1.0000000       NA
form1  QG3    G    3 1.0000000  0.00000000  2.0000000       NA
form1  QG4    G    3 0.7000000 -3.00000000  1.0000000       NA
form1  QG5    G    3 0.7000000 -2.00000000  2.0000000       NA
form1  QG6    G    3 0.7000000 -1.00000000  3.0000000       NA
form1  QG8    G    4 0.7000000 -2.00000000  0.0000000 2.000000
form1  QP1    P    3 0.7344000 -2.11500000  0.0530000       NA
form1  QP2    P    3 0.8020000 -0.99850000  0.9985000       NA
form1  QP3    P    3 0.8237000  0.07900000  1.8952000       NA
form1  QP4    P    3 0.5998000 -3.19340000  1.1550000       NA
form1  QP5    P    3 0.6258000 -2.10710000  2.1071000       NA
form1  QP6    P    3 0.6319000 -1.05120000  3.0354000       NA
form1  QP7    P    4 0.6483000 -2.33070000  0.2053000 2.125400
form2  QB1   B2    2 0.9736860 -2.50974572  0.0000000       NA
form2  QB3   B3    2 0.7560111  0.25955944  0.2991986       NA
form2  QG1    G    3 0.7700730 -3.85306557 -1.1357144       NA
form2  QG3    G    3 1.0288158 -1.12433239  1.4054329       NA
form2  QG5    G    3 0.8033210 -3.74832196  1.5112735       NA
form2  QG7    G    4 0.9167175 -3.81423155 -1.1428328 1.613180
form2  QP1    P    3 0.7256544 -4.10048328 -1.2231697       NA
form2  QP3    P    3 0.6181345 -1.06795481  1.0809555       NA
form2  QP5    P    3 0.5632958 -3.93255199  1.3176705       NA
form2  QP7    P    4 0.5277231 -4.28903421 -0.7262809 1.149967
form2  QB2   B2    2 0.9845137 -1.15960242  0.3000000       NA
form2  QG2    G    3 0.9434385 -2.55422669  0.2331392       NA
form3  QB2   B2    2 1.5268054  1.01997505  0.3000000       NA
form3  QB4   B3    2 1.3534871  1.58018603  0.4643032       NA
form3  QG2    G    3 1.5803144  0.16910510  1.5452752       NA
form3  QG4    G    3 1.1897892 -1.44156820  1.6082322       NA
form3  QG6    G    3 0.9773134  0.18749580  3.2815727       NA
form3  QG8    G    4 1.2314105 -0.73990378  1.0218542 2.315625
form3  QP2    P    3 1.1810182  0.04046525  1.5351101       NA
form3  QP4    P    3 0.9534385 -1.49477547  1.6427269       NA
form3  QP6    P    3 1.0670629  0.07743022  3.4615809       NA
form3  QP8    P    4 0.6973531 -0.56943438  0.9358040 2.118880
form3  QB1   B2    2 1.6124478  0.09865741  0.0000000       NA
}
}
\keyword{datasets}
