% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{paramS1}
\alias{paramS1}
\title{Small Item Parameter Data Frame # 1. (3 items)\cr}
\format{
A data frame with 4 rows and 7 variables
\describe{
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{p1}{num: First item parameter: slope}
 \item{p2}{num: 2nd  item parameter}
 \item{p3}{num: 3rd item parameter}
 \item{p4}{num: 4th item parameter}
}
}
\source{
Generated by author.
}
\usage{
paramS1
}
\description{
Small Item Parameter Data Frame # 1. (3 items)\cr
}
\details{
\preformatted{
name type ncat  p1  p2   p3   p4
  Q1    B    2 0.9  0.0  NA   NA
  Q2   B3    2 1.0  0.0 0.2   NA
  Q3    G    4 1.0 -2.0 0.0  2.0
  Q4    P    4 0.8 -2.0 0.0  2.0
  }
   item type = "B" or "B3" or "G" or "P" \cr
   ncat is the max # of categories \cr
}
\keyword{datasets}
