% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{paramS2}
\alias{paramS2}
\title{Small Item Parameter Data Frame # 2. (8 items)\cr}
\format{
A data frame with 8 rows and 7 variables
\describe{
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{p1}{num: First item parameter: slope}
 \item{p2}{num: 2nd  item parameter}
 \item{p3}{num: 3rd item parameter}
 \item{p4}{num: 4th item parameter}
}
}
\source{
Generated by author.
}
\usage{
paramS2
}
\description{
Small Item Parameter Data Frame # 2. (8 items)\cr
}
\details{
\preformatted{
 name type ncat  p1  p2   p3 p4
  Q11    B    2 0.9  0.0  NA NA
  Q12    B    2 0.9  1.0  NA NA
  Q21   B3    2 1.0  0.0 0.2 NA
  Q22   B3    2 1.0  1.0 0.2 NA
  Q31    G    3 1.0 -1.0 1.0 NA
  Q32    G    4 1.0 -2.0 0.0  2.0
  Q41    P    3 0.8 -1.0 1.0 NA
  Q42    P    4 0.8 -2.0 0.0  2.0
  }
   item type = "B" or "B3" or "G" or "P" \cr
   ncat is the max # of categories \cr
}
\keyword{datasets}
