% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{paramS3}
\alias{paramS3}
\title{Small Item Parameter Data Frame # 3.  (20 items)\cr}
\format{
A data frame with 20 rows and 7 variables
\describe{
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{p1}{num: First item parameter: slope}
 \item{p2}{num: 2nd  item parameter}
 \item{p3}{num: 3rd item parameter}
 \item{p4}{num: 4th item parameter}
}
}
\source{
Generated by author.
}
\usage{
paramS3
}
\description{
Small Item Parameter Data Frame # 3.  (20 items)\cr
}
\details{
\preformatted{
name  type ncat p1        p2         p3         p4
QB1   B   2    1         -1           NA        NA
QB2   B2  2    1          0           NA        NA
QB3   B2  2    1          1           NA        NA
QB4   B3  2    .8         1           .2        NA
QG1   G   3    1.0       -2          0          NA
QG2   G   3    1.0       -1          1          NA
QG3   G   3    1.0        0          2          NA
QG4   G   3    0.7       -3          1          NA
QG5   G   3    0.7       -2          2          NA
QG6   G   3    0.7       -1          3          NA
QG7   G   4    1.0       -2          0           2
QG8   G   4    0.7       -2          0           2
QPN1  PN  3    0.7344    -2.1150    -1.0310     NA
QPN2  PN  3    0.8020    -0.9985     0.0000     NA
QPN3  PN  3    0.8237     0.0790     0.9871     NA
QPN4  PN  3    0.5998    -3.1934    -1.0192     NA
QPN5  PN  3    0.6258    -2.1071     0.0000     NA
QPN6  PN  3    0.6319    -1.0512     0.9921     NA
QPN7  PN  4    0.6483    -2.3307    -1.0627      0
QPN8  PN  4    0.4231    -1.9974    -1.0154      0
  }
   item type = "B" or "B3" or "G" or "PN" \cr
   ncat is the max # of categories \cr
}
\keyword{datasets}
