% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfuncs.R
\name{read.param}
\alias{read.param}
\title{Reading Parameter File}
\usage{
read.param(
  infile,
  skip = 0,
  nrows = -1,
  na.strings = "NA",
  sep = "",
  print = 0
)
}
\arguments{
\item{infile}{Input file name}

\item{skip}{# of lies to skip}

\item{nrows}{NOT used}

\item{na.strings}{A character indicating missing value}

\item{sep}{Delimite character}

\item{print}{= 1 to print result}
}
\description{
Reading Parameter File
}
\details{
It is assumed that the data file has the following quantities
  in the following order.\cr
\cr
  item_name, item_type, ncat, p1, p2, ..., p[ncat+1]\cr
\cr
The first line must consist of the variable names shown above. \cr
If the data file has header lines berore variable names, those lines
must be skipped by specifying skip= parameter.
\cr
  When item_type == "B", ( ncat == 2 )\cr
    p1 is the discrimination parameter\cr
    p2 is the difficulty parameter\cr
    p3 is the gussing parameter\cr
\cr
  When item_type == "G",\cr
    p1 is the discrimination parameter\cr
    p2 is the category threshold parameter b_{j1}\cr
    p3 is the category threshold parameter b_{j2}\cr
\cr
    p_ncat is the category threshold parameter b_{j,[ncat-1]}\cr
\cr
  When item_type == "P",\cr
    p1 is the discrimination parameter\cr
    p2 is the step parameter b_{j1}\cr
    p3 is the step parameter b_{j2}\cr
\cr
    p_ncat is the step parameter b_{j,[ncat-1]}\cr
\cr
  When item_type == "PN,\cr
    p1 is the slope parameter\cr
    p2 is the intercept parameter b_{j1}\cr
    p3 is the intercept parameter b_{j2}\cr
\cr
    p_ncat is the intercept parameter b_{j,[ncat-1]}\cr
\cr
  When item_type == "N,\cr
    p1 - p_(ncat[j]-1) are the slope parameters\cr
    p_ncat[j] - p_2*(ncat[j]-1) are the intercept parameters\cr
\cr
  Regardless of the item_type, there will be ncat[j]\cr
  item paramters for item j, except for the Binary Items\cr
  which have the gusseing param as the ncat[j]+1st,\cr
  and the Nominal Items which have 2*(ncat[j]-1) item parameters.\cr
\cr
}
