% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfuncs.R
\name{read.weight}
\alias{read.weight}
\title{Reading Weight File}
\usage{
read.weight(
  infile,
  skip = 0,
  nrows = -1,
  na.strings = ".",
  sep = "",
  print = 0
)
}
\arguments{
\item{infile}{Input file name}

\item{skip}{# of lies to skip}

\item{nrows}{NOT used}

\item{na.strings}{A character indicating missing value}

\item{sep}{Delimite character}

\item{print}{= 1 to print result}
}
\description{
Reading Weight File
}
\details{
It is assumed that the data file has the following quantities
  in the following order.
\cr
  item_name item_type  ncat  w  v0  v1  v2  ..., v[ncat-1] \cr
\cr
 where w is the item weight and v0,v1..., are the category weight. \cr
}
