% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_irt.R
\name{rel_irt}
\alias{rel_irt}
\title{Calculation of Test Reliability and SEM under IRT}
\usage{
rel_irt(
  param,
  weight = NULL,
  maxscore = 100,
  npoints = 51,
  thmin = -4,
  thmax = 4,
  thdist = 1,
  print = 1,
  plot = 1
)
}
\arguments{
\item{param}{Item Parameter Data Frame}

\item{weight}{Weight data frame}

\item{maxscore}{The maximum score to be used to calculate var(X) etc.}

\item{npoints}{# of discrete points for theta}

\item{thmin}{Minimum value of discrete thata value}

\item{thmax}{Maximum value of discrete thata value}

\item{thdist}{Type of theta distribution \cr
= 0 to use uniform,  \cr = 1 to use N(0,1)}

\item{print}{= 1 to print result}

\item{plot}{= 1 or 2 to plot result}
}
\value{
A list of: \cr\cr
thmin, thmax, npoints, thdist \cr
SigmaX2 Observed Score Variance \cr
SigmaT2 True Score Variance \cr
SigmaT22 Shoulld be the same as above \cr
SigmaE2 Error Score Variance \cr
aSEM Average Standard Error of Measurement \cr
rel The reliability coefficient \cr
\cr
SigmaX_theta2 Variance of Thetahat \cr
SigmaT_theta2 Variance of Theta\cr
SigmaE_theta2 Average Variance of Thetahat \cr
aSEM_theta Average Standard Error of Measurement of Thetahat \cr
rel_theta The reliability coefficient of Thetahat based Score \cr
\cr
SigmaX_xs2 Variance of X in [0,maxscore] \cr
SigmaT_xs2 Variance of T in [0,maxscore] \cr
SigmaE_xs2 Average Variance of E in [0,maxscore]  \cr
aSEM_xs Average Standard Error of Measurement of X in [0,maxscore] \cr
rel_xs The reliability coefficient \cr
\cr
SigmaX_ys2 Variance of Thetahat in [0,maxscore] \cr
SigmaT_ys2 Variance of Theta in [0,maxscore] \cr
SigmaE_ys2 Average Variance of Thetahat given theta in [0,maxscore]  \cr
aSEM_ys Average Standard Error of Measurement of Thetahat in [0,maxscore]
 \cr
rel_ys The reliability coefficient \cr
\cr
 SEM_x      SEM of X at True Score \cr
 SEM_t      SEM of X and Y(thetahat) at Theta \cr
 SEM_x100   Rescaled SEM of X and Y(thetahat) at Rescaled True Score \cr
 SEM_t100   Rescaled SEM of X and Y(thetahat) at theta \cr
}
\description{
Calculation of Test Reliability and SEM under IRT
}
\details{
Test reliability of a text Z is defined as:
\preformatted{
 rel(Z) = var(T) / var(Z) = 1 - var(E)/var(Z)
}
where T and E are, resp, the true score and the error score of Z. \cr
For observed score: X = trf(theta) + E   and  E(E|theta) = 0,
\preformatted{
var(E) = E(var(E|theta)) = E(var(X|theta))
var(X) = E(var(X|theta)) + var(E(X|theta)) = var(E) + var(trf)
var(T) = var(trf)
}

For thetahat based score: Y = thetahat = theta + error and E(error|theta)=0,
\preformatted{
var(error) = E(var(error|theta)) = var(Y|theta) = 1/info(theta)
var(Y) = E(var(Y|theta)) + var(E(Y|theta)) = E(1/info(theta)) + var(theta)
var(theta) = 1 or var(uniform)
}
\cr
The standard error of measurement, SEM, is defined as, \cr
For X
\preformatted{
 SEMx(X) = sqrt( var(E|X) )  or  SEMx(theta) = sqrt( var(E|trf(theta)) )
}
For Y,
\preformatted{
 SEMt(Y) = sqrt( var(error|theta) ) or
 SEMt(X) = sqrt( var(error|trf(theta)) )
}
Their averages are denoted, resp, as \cr
aSEM or aSEM_theta.
\cr
Note that, SEMx(X) and SEMx(theta) are convex and
SEMt(Y) and SEMt(theta) are concave. \cr

Note that the value depends on the range and the shape
of the theta distribution.
\cr
Try changing npoints, thmin and thmax.
}
\examples{
res <- rel_irt( paramA1, thmin=-4, thmax=4, plot=2 )
res <- rel_irt( paramA1, thmin=-4, thmax=4, thdist=0, plot=2 )

}
\references{
Lord, F. M. (1980).
Applications of Item Response Theory To Practical Testing Problems.
Lawrence Erlbaum Associates. Mahwah, New Jergey


Samejima, F. (1994) Estimation of Reliability Coefficients Using
the Test Information Function and Its Modifications.
Applied Psychological Measurement. 18, 229-244.

Toyoda. H. (1989) The Methods for Estimating the Reliability Coefficient
under Item Response Model. (in Japanese)
Japanese Journal of Educational Psychology. vol. 37, 283-285.
}
