% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smn.R
\name{smn}
\alias{smn}
\title{Parameter Estimation of Parametric (Normal pdf)
Scored Multinomial Distributions}
\usage{
smn(
  x,
  f,
  mu = NULL,
  sigma = NULL,
  estmu = 1,
  estsigma = 1,
  maxiter = 100,
  eps = 0.00000001,
  print = 1,
  plot = 0
)
}
\arguments{
\item{x}{Score (numerical category label)}

\item{f}{Frequency vector which sums to unity}

\item{mu}{Initial value of the mean paramter to be estimated or NULL}

\item{sigma}{Initial value of sigma the sigma parameter to be estimated
or NULL}

\item{estmu}{= 0 to skip the estimation of mu}

\item{estsigma}{= 0 to skip the estimation of sigma}

\item{maxiter}{max # of iterations}

\item{eps}{eps for convergence}

\item{print}{= 1 to print result}

\item{plot}{= 1 to plot result}
}
\value{
A list of: \cr
P A vector of estimated probability\cr
x A vector of score\cr
mu Estimated mean parameter\cr
sigma Estimated standard deviation parameter\cr
estmu = 1 to estimate mu\cr
estsigma = 1 to estimate sigma\cr
llh Maximized log likelihood\cr
mean Sample mean\cr
std Sample standard deviation
}
\description{
Maximum likelihood estimation of the parameters of
the parametric (normal pdf) scored multinomial distribution. \cr
Japanese help file: (\link[lazy.irt]{smn_JPH})
}
\details{
Given data, (x,f), where x is the domain and f is the associated
frequency, \cr
this program maximizes the following multinomial log likelihood \cr
  \eqn{ lnL = \sum_{i} f_i \log{p(x_i)} } \cr
where \cr
 \eqn{ p(x_i) = c \times dnorm( x_i, mu, sigma )  } \cr
 and  \eqn{ c = 1 / \sum_{i} dnorm( x_i, mu, sigma ) }  \cr
with respect to mu and sigma.
}
\examples{
set.seed(1701)
n <- 1000
npoints <- 21
resg <- list(midpoints=c(-2,-1,0,1,2), freq=c(5,6,8,6,2))
res <- smn( resg$midpoints, resg$freq, print=1, plot=1, mu=0, sigma=1 )

}
