% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsmnw.R
\name{sumsmnw12}
\alias{sumsmnw12}
\title{Distribution of the Weighted Sum of Two Independent
Scored Multinomial Distributions}
\usage{
sumsmnw12(p1, p2, w2 = 1, eps = 0.00000001, print = 0)
}
\arguments{
\item{p1}{Probability of the fist random variable}

\item{p2}{Probability of the second random variable}

\item{w2}{Integer weight to the second random variable}

\item{eps}{small number}

\item{print}{= 1 to print result}
}
\value{
A matrix of (score, prob)
}
\description{
Distribution of the Weighted Sum of Two Independent
Scored Multinomial Distributions
}
\details{
This program calculats the distrobution of\cr
     X12 = X1 + w2 X2\cr
    where\cr
     Xi  is distributed as Scored Multinomial with  pi,  i=1,2.\cr
   That is, this program calculates the probability that\cr
     p12[a] = Pr( X12 = a ) , a=0,1,..., n12=(n1-1)+w2*(n2-1)\cr
\cr
 Note\cr
   length(pi) = ni = #'  of categories of the i-th r.v.\cr
              = mi + 1   where mi is the max score of the i-th r.v.\cr
\cr
}
\examples{
sumsmnw12( 1:3, 1:2, 2, print=1 )

}
\references{
Mayekawa, S., & Arai, S. (2008).
Distribution of the Sum of Scored Multinomial Random Variables
and Its Application to the Item Response Theory.
In K. Shigemasu, A. Okada, T.Imaizumi, & T. Hoshino (Eds.)
New Trends in Psychometrics. Tokyo: University Academic Press.
}
