% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uLRT.R
\name{uLRT}
\alias{uLRT}
\title{Item Parameter Estimation of Unidimensional LRT}
\usage{
uLRT(
  Uc,
  U = NULL,
  groupvar = NULL,
  idvar = NULL,
  ncat = NULL,
  type = NULL,
  itemname = NULL,
  nclass = 5,
  V = NULL,
  fixeditems = NULL,
  baseform = 1,
  monotone = 0,
  rho = NULL,
  estrho = 0,
  alpha = rep(1, nclass),
  vmin = 0.0001,
  vmax = 1 - vmin,
  maxiter = 200,
  eps = 0.0000001,
  epsd = 0.000001,
  maxiter2 = 1,
  eps2 = 0,
  nstrict = 9,
  maxiter22 = 5,
  minp1 = 0.1,
  maxabsparam = 20,
  SQUAREM = 3,
  nSQUAREM = 1,
  minalpha = -999,
  maxalpha = -1,
  always = 1,
  reset1 = 0,
  reset2 = 1,
  print = 2,
  plot = 0,
  smallP = 0,
  debug = 0
)
}
\arguments{
\item{Uc}{n x nitems+1 or +2    compressed item resopnse data frame
in BILOG-MG's expanded format. (idvar and groupvar)}

\item{U}{n x sum(ncat)+1 or +2 uncompressed irem response data frame}

\item{groupvar}{name of the grouping varible contained as a column of Uc
or NULL.}

\item{idvar}{name of the id varible contained as a column of Uc
or NULL}

\item{ncat}{nitems x 1 #  of categories for each item or fNULL.}

\item{type}{nitems x 1 vector of item types consisting of:
"Bn" | "G" | "PN" | "P" \cr
If NULL, P will be used. \cr
*** Currently, polytomous items are not allowed.}

\item{itemname}{nitems x 1 vector of item names or NULL.}

\item{nclass}{# of classes}

\item{V}{npoitns x sum(ncat)  initial value for item param\cr
Only the subset of the items can be given.\cr
param$name and the colname(Uc) will be used to match the items.}

\item{fixeditems}{list of items whose item parameters are to be fixed\cr
 to the values given in the param data frame.\cr
Item numbers or item names\cr
*** Currently, this option is not available.}

\item{baseform}{base form number whose theta distribution is
fixed at the values given in msn[baseform,] \cr
 If there are fixed items, the values of the item paramters
 given in the param data frame must be on the baseform scale.\cr}

\item{monotone}{= 1 to enforce monotonicity}

\item{rho}{initial value of probability vector of each latent class.}

\item{estrho}{= 1  to estimate multigroup theta means}

\item{alpha}{= vector of Dirichle parameters of length nclass \cr
or a number A to set alpha=rep(A,nclass) \cr
or a negative number -A to set alpha=A*normal pdf}

\item{vmin}{minimum value of V}

\item{vmax}{maximum value of V}

\item{maxiter}{max # of iterations}

\item{eps}{eps for the relative improvement of lmlh}

\item{epsd}{eps for the max. abs. diff. of msn\cr
Seems this is important.}

\item{maxiter2}{max # of iterations for ordinal_reg and smn
when llll <= nstrict.}

\item{eps2}{eps for the relative improvement of
llh in ordinal_reg and smn.}

\item{nstrict}{maxiter2 will be reduced to maxiter22
after nstrict iterations.}

\item{maxiter22}{max # of iterations for ordinal_reg
and smn when llll > nstrict.}

\item{minp1}{Minimum value of p1 parameter for type != "N" items.}

\item{maxabsparam}{Maximum absolute value of parameters.}

\item{SQUAREM}{= 3 :   See the help of iSQUAREM in lazy.accel package.}

\item{nSQUAREM}{when to star iSQUAREM}

\item{minalpha}{= -999 :   See the help of iSQUAREM in lazy.accel package.}

\item{maxalpha}{= -1 :   See the help of iSQUAREM in lazy.accel package.}

\item{always}{= 0 :   See the help of iSQUAREM in lazy.accel package.}

\item{reset1}{= 1 :   See the help of iSQUAREM in lazy.accel package.}

\item{reset2}{= 2 :   See the help of iSQUAREM in lazy.accel package.}

\item{print}{= 1 to print the result}

\item{plot}{= 1 to plot the estimated theta distributions}

\item{smallP}{= Minimum value of P}

\item{debug}{= 1 to print intermediate result}
}
\value{
A list of:
V Estimated Item Parameters in a data frame \cr
rho Class distribution \cr
converged = 1 if converged, = 0 otherwise. \cr
lmlh Log Marginal LIkelihood maximized \cr
iter_hist A list of iteration history \cr
H Posterior distribution of theta given U.  \cr
id Id variable \cr
}
\description{
Marginal likelihood will be maximized by the EM algorithm
using ordinal_reg funcion and smn function in the M-step.
}
\details{
*** Currently, polytomous items are not allowed. ****** \cr
*** Currently, multi-group analysis is not available. \cr
\cr
 Uc is the compressed data if it is n x nitems. \cr
 U is the uncompressed data if it is n x sum(ncat). \cr
 \cr
 When U, instead of Uc, is given: \cr
 Item names come from itemname or paste("Q",1:nitems,sep=""). \cr
 ncat must be given. \cr

 When Uc is given: \cr
  Item name comes from colnames(Uc) or itemname. \cr
  ncat can be calculated from Uc.\cr

 \cr
 Note on the iSQUAREM: \cr
 Try always=1 with maxalpha=-1 or less first. \cr
 Changing to reset1=1 and reset2=2 or increasing nSQUAREM may help. \cr
 If it seems not working, use always=0 with maxalpha=1 or less. \cr
 Changing to reset1=1 and reset2=2 may help. \cr
 If all of the above fail, be patient and use SQUAREM=0.
 \cr
}
\examples{
#
#### In the following examples, maxiter is set to 20 which is
#### not large enough to obtain convergence.
####
#
#
#
set.seed(1701)

param=paramB1[c(1:3,7:9,13:15),]
thmin=-2; thmax=2; npoint=5
N=1000

# discrete theta
# theta0=seq(thmin,thmax,length=npoint)
theta0=c(-2, -1, 0, 2, 3)
theta=unlist(lapply( theta0, rep, round(N/npoint) ))
res2 <- gendataIRT( 1, paramB1, theta=theta, compress=1 )
Uc=as.data.frame(res2$U)
ncat=res2$ncat
type=res2$type

nclass=5
res1 <- uLRT( Uc, nclass=nclass, estrho=1, monotone=1, alpha=20
, maxiter=20, plot=1, print=1 )

# normal theta
Uc2 <- gendataIRT( 1, paramB1, npoints=N, thdist="rnorm", compress=1 )$U
Uc2 <- as.data.frame(Uc2)

nclass=5
res1 <- uLRT( Uc2, nclass=nclass, estrho=1, monotone=1
, maxiter=20, plot=1, print=1 )

}
\references{
Varadhan, R. and Roland, C.(2007) Simple and Globally Convergent Methods
for Accelerating the Convergence of Any EM Algorithm.
Scandinavian Journal of Statistics, Vol. 35: 335-353. \cr
Shojima, K. (2008a) Neural test theory.
In K. Shigemasu, A. Okada, T. Imaizumi, & T. Hoshino (Eds.)
New trends in psychometrics. Universal Academy Press, Inc
}
