% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{weightA1}
\alias{weightA1}
\title{Item Category Weight Data Frame # 1. (10 items of all types)\cr}
\format{
A data frame with 7 rows and 6 variables.
\describe{
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{w}{num: item weight}
 \item{v0}{num: category weight for the zero-th category}
 \item{v1}{num: category weight for the 1st category}
 \item{v2}{num: category weight for the 2nd category}
 \item{v3}{num: category weight for the 3rd category}
}
}
\source{
Generated by author.
}
\usage{
weightA1
}
\description{
Item Category Weight Data Frame # 1. (10 items of all types)\cr
}
\details{
\preformatted{
  name type  ncat w  v0 v1 v2 v3
     Q1    B    2 1  0  1 NA NA
     Q2   B3    2 1  0  1 NA NA
     Q3   Bn    2 1  0  1 NA NA
     Q4  Bn3    2 1  0  1 NA NA
     Q5    P    3 1  0  1  2 NA
     Q6    G    3 1  0  1  2 NA
     Q7   Gn    3 1  0  1  2 NA
     Q8    P    4 1  0  1  2  3
     Q9    G    4 1  0  1  2  3
    Q10   Gn    4 1  0  1  2  3
  }
   item type = all types \cr
   ncat is the max # of categories \cr
 \cr\cr
 The type P item was converted to type G and Gn items using
 fitG2P_ls and fitGn2P_ls functions, resp, so that
 the icrfs of type G and Gn items are as close as
 the irf of type P item.
}
\keyword{datasets}
