% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc2.R
\docType{data}
\name{weightB14}
\alias{weightB14}
\title{Item Weight Data Frame for paramB1 or paramB2 5 (18 items)\cr}
\format{
A data frame with 18 rows and 6 variables
\describe{
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{w}{num: item weight}
 \item{v0}{num: item category weight for 0}
 \item{v1}{num: item category weight for 1}
}
}
\source{
Generated by author.
}
\usage{
weightB14
}
\description{
Item Weight Data Frame for paramB1 or paramB2 5 (18 items)\cr
}
\details{
\preformatted{
name type ncat  w   v0 v
 Q1    B    2   2   0  1
 Q2    B    2   2   0  1
 Q3    B    2   2   0  1
 Q4    B    2   2   0  1
 Q5    B    2   2   0  1
 Q6    B    2   2   0  1
 Q7    B    2   1   0  1
 Q8    B    2   1   0  1
 Q9    B    2   1   0  1
Q10    B    2   1   0  1
Q11    B    2   1   0  1
Q12    B    2   1   0  1
Q13    B    2   3   0  1
Q14    B    2   3   0  1
Q15    B    2   3   0  1
Q16    B    2   3   0  1
Q17    B    2   3   0  1
Q18    B    2   3   0  1
  }
   \cr
   item type = "B" \cr
   ncat is the max # of categories \cr\cr
   Large weight for low discriminative items.
}
\keyword{datasets}
