% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{weightS12}
\alias{weightS12}
\title{Small Item Weight Data Frame # 3.\cr
To be used in conjunction with paramS1.}
\format{
A data frame with 4 rows and 8 variables
\describe{
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{w}{num: item weight}
 \item{v0}{num: category weight for the zero-th category}
 \item{v1}{num: category weight for the 1st category}
 \item{v2}{num: category weight for the 2nd category}
 \item{v3}{num: category weight for the 3rd category}
}
}
\source{
Generated by author.
}
\usage{
weightS12
}
\description{
Small Item Weight Data Frame # 3.\cr
To be used in conjunction with paramS1.
}
\details{
\preformatted{
name type ncat  w    v0   v1   v2   v3
Q1    B    2    1    0    1 ,,
Q2    B3   2    1    0    2 ,,
Q3    G    4    1    0    1    2    5
Q4    P    4    1    0    0    0    1
}
   item type = "B" or "B3" or "G" or "P" \cr
   ncat is the max # of categories \cr
   w is the item weight. (all equal)\cr
   v's are the category weight
}
\keyword{datasets}
