% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{weightS21}
\alias{weightS21}
\title{Small Item Weight Data Frame # 5.\cr
To be used in conjunction with paramS2.}
\format{
A data frame with 8 rows and 8 variables
\describe{
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{w}{num: item weight}
 \item{v0}{num: category weight for the zero-th category}
 \item{v1}{num: category weight for the 1st category}
 \item{v2}{num: category weight for the 2nd category}
 \item{v3}{num: category weight for the 3rd category}
}
}
\source{
Generated by author.
}
\usage{
weightS21
}
\description{
Small Item Weight Data Frame # 5.\cr
To be used in conjunction with paramS2.
}
\details{
\preformatted{
name type ncat  w    v0   v1   v2   v3
Q11    B    2   1    0    1
Q12    B    2   2    0    1
Q21    B3   2   1    0    1
Q22    B3   2   2    0    1
Q31    G    3   1    0    1    2
Q32    G    4   2    0    1    2    3
Q41    P    3   1    0    1    2
Q42    P    4   2    0    1    2    3";
}
   item type = "B" or "B3" or "G" or "P" \cr
   ncat is the max # of categories \cr
   w is the item weight. \cr
   v's are the category weight. (all natural)
}
\keyword{datasets}
