% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdatadoc.R
\docType{data}
\name{weightS3}
\alias{weightS3}
\title{Small Item Weight Data Frame #7
To be used in conjunction with paramS3\cr}
\format{
A data frame with 20 rows and 8 variables
\describe{
 \item{name}{character:  Item name}
 \item{type}{character:  Item type}
 \item{ncat}{num: # of categories: 2 for binary items}
 \item{w}{num: item weight}
 \item{v0}{num: category weight for the zero-th category}
 \item{v1}{num: category weight for the 1st category}
 \item{v2}{num: category weight for the 2nd category}
 \item{v3}{num: category weight for the 3rd category}
}
}
\source{
Generated by author.
}
\usage{
weightS3
}
\description{
Small Item Weight Data Frame #7
To be used in conjunction with paramS3\cr
}
\details{
\preformatted{
name  type ncat w   v0   v1   v2    v3
QB1   B   2     1     0   1   NA    NA
QB2   B2  2     1     0   1   NA    NA
QB3   B2  2     1     0   1   NA    NA
QB4   B3  2     1     0   1   2     NA
QG1   G   3     1     0   1   2     NA
QG2   G   3     1     0   1   2     NA
QG3   G   3     1     0   1   2     NA
QG4   G   3     1     0   1   2     NA
QG5   G   3     1     0   1   2     NA
QG6   G   3     1     0   1   2     NA
QG7   G   4     1     0   1   2      3
QG8   G   4     1     0   1   2      3
QPN1  PN  3     1     0   1   2     NA
QPN2  PN  3     1     0   1   2     NA
QPN3  PN  3     1     0   1   2     NA
QPN4  PN  3     1     0   1   2     NA
QPN5  PN  3     1     0   1   2     NA
QPN6  PN  3     1     0   1   2     NA
QPN7  PN  4     1     0   1   2      3
QPN8  PN  4     1     0   1   2      3
}
   item type = "B" or "B3" or "G" or "PN" \cr
   ncat is the max # of categories \cr
   w is the item weight. (all equal)\cr
   v's are the category weight
}
\keyword{datasets}
